feat(tool): localize http status code descriptions
