+++ tools/http-status-code-lookup/descriptions/ar.json
@@ -1,0 +1,52 @@
+{
+  "100": "استلم الخادم رؤوس الطلب ويجب على العميل المتابعة لإرسال جسم الطلب.",
+  "101": "طلب المستعلم من الخادم تبديل البروتوكولات ووافق الخادم.",
+  "102": "استلم الخادم الطلب وهو يعالجه، لكن لا يوجد رد متاح بعد.",
+  "103": "يستخدم لإرجاع بعض رؤوس الاستجابة قبل رسالة HTTP النهائية.",
+  "200": "نجح الطلب. معنى النجاح يعتمد على طريقة HTTP.",
+  "201": "تم تنفيذ الطلب وأدى إلى إنشاء مورد جديد.",
+  "202": "تم قبول الطلب للمعالجة، لكن المعالجة لم تكتمل.",
+  "204": "عالج الخادم الطلب بنجاح ولا يعيد أي محتوى.",
+  "206": "يسلم الخادم جزءًا فقط من المورد بسبب رأس النطاق المرسل من العميل.",
+  "300": "الطلب له أكثر من رد محتمل. يجب على وكيل المستخدم اختيار واحد منها.",
+  "301": "تم تغيير عنوان URL للمورد المطلوب بشكل دائم.",
+  "302": "تم تغيير عنوان URL للمورد المطلوب مؤقتًا.",
+  "304": "يشير إلى أن المورد لم يتم تعديله منذ الإصدار المحدد بواسطة رؤوس الطلب.",
+  "307": "يعيد الخادم توجيه وكيل المستخدم إلى مورد مختلف، كما هو موضح برأس Location.",
+  "308": "المورد موجود الآن بشكل دائم في عنوان URL آخر، محدد برأس Location.",
+  "400": "لا يمكن للخادم أو لن يعالج الطلب بسبب خطأ واضح من العميل.",
+  "401": "المصادقة مطلوبة وفشلت أو لم يتم توفيرها بعد.",
+  "403": "الطلب كان صالحًا، لكن الخادم يرفض الإجراء.",
+  "404": "لم يتم العثور على المورد المطلوب لكن قد يكون متاحًا في المستقبل.",
+  "405": "طريقة الطلب غير مدعومة للمورد المطلوب.",
+  "406": "المورد المطلوب قادر على توليد محتوى غير مقبول فقط وفقًا لرؤوس Accept المرسلة في الطلب.",
+  "408": "انتهت مهلة الخادم في انتظار الطلب.",
+  "409": "لا يمكن معالجة الطلب بسبب تعارض في الطلب.",
+  "410": "المورد المطلوب لم يعد متاحًا ولن يكون متاحًا مرة أخرى.",
+  "411": "لم يحدد الطلب طول محتواه، وهو مطلوب من المورد المطلوب.",
+  "412": "الخادم لا يلبي أحد الشروط المسبقة التي وضعها المستعلم على الطلب.",
+  "413": "الطلب أكبر مما يرغب الخادم أو قادر على معالجته.",
+  "414": "كان URI المقدم طويلًا جدًا للخادم لمعالجته.",
+  "415": "كيان الطلب له نوع وسائط لا يدعمه الخادم أو المورد.",
+  "416": "طلب العميل جزءًا من الملف، لكن الخادم لا يمكنه توفير ذلك الجزء.",
+  "417": "لا يمكن للخادم تلبية متطلبات حقل Expect لرأس الطلب.",
+  "418": "تم تعريف هذا الرمز في عام 1998 كواحد من نكات كذبة أبريل التقليدية لـ IETF.",
+  "422": "كان الطلب منسقًا جيدًا لكن لم يمكن اتباعه بسبب أخطاء دلالية.",
+  "425": "يشير إلى أن الخادم غير مستعد للمخاطرة بمعالجة طلب قد يتم إعادة تشغيله.",
+  "426": "يجب على العميل التبديل إلى بروتوكول مختلف مثل TLS/1.0.",
+  "428": "يتطلب الخادم الأصلي أن يكون الطلب مشروطًا.",
+  "429": "أرسل المستخدم طلبات كثيرة جدًا في فترة زمنية معينة.",
+  "431": "الخادم غير مستعد لمعالجة الطلب لأن حقول الرأس كبيرة جدًا.",
+  "451": "الخادم يرفض الوصول إلى المورد نتيجة لطلب قانوني.",
+  "500": "رسالة خطأ عامة، تُعطى عند مواجهة حالة غير متوقعة.",
+  "501": "الخادم إما لا يتعرف على طريقة الطلب، أو يفتقر إلى القدرة على تنفيذ الطلب.",
+  "502": "كان الخادم يعمل كبوابة أو وكيل وتلقى ردًا غير صالح من الخادم المنبع.",
+  "503": "الخادم غير متاح حاليًا (لأنه مثقل أو متوقف للصيانة).",
+  "504": "كان الخادم يعمل كبوابة أو وكيل ولم يتلق ردًا في الوقت المناسب من الخادم المنبع.",
+  "505": "الخادم لا يدعم إصدار بروتوكول HTTP المستخدم في الطلب.",
+  "506": "التفاوض الشفاف للمحتوى للطلب يؤدي إلى مرجع دائري.",
+  "507": "الخادم غير قادر على تخزين التمثيل المطلوب لإكمال الطلب.",
+  "508": "اكتشف الخادم حلقة لا نهائية أثناء معالجة الطلب.",
+  "510": "مطلوب امتدادات إضافية للطلب حتى يتمكن الخادم من تنفيذه.",
+  "511": "يحتاج العميل إلى المصادقة للحصول على وصول الشبكة."
+}
