+++ tools/http-status-code-lookup/descriptions/de.json
@@ -1,0 +1,52 @@
+{
+  "100": "Der Server hat die Anfrage-Header erhalten und der Client sollte mit dem Senden des Anfragekorpers fortfahren.",
+  "101": "Der Anfragende hat den Server gebeten, die Protokolle zu wechseln, und der Server hat zugestimmt.",
+  "102": "Der Server hat die Anfrage erhalten und verarbeitet sie, aber es ist noch keine Antwort verfugbar.",
+  "103": "Wird verwendet, um einige Antwort-Header vor der endgultigen HTTP-Nachricht zuruckzugeben.",
+  "200": "Die Anfrage war erfolgreich. Die Bedeutung des Erfolgs hangt von der HTTP-Methode ab.",
+  "201": "Die Anfrage wurde erfullt und fuhrte zur Erstellung einer neuen Ressource.",
+  "202": "Die Anfrage wurde zur Verarbeitung akzeptiert, aber die Verarbeitung ist noch nicht abgeschlossen.",
+  "204": "Der Server hat die Anfrage erfolgreich verarbeitet und gibt keinen Inhalt zuruck.",
+  "206": "Der Server liefert nur einen Teil der Ressource aufgrund eines vom Client gesendeten Bereichs-Headers.",
+  "300": "Die Anfrage hat mehr als eine mogliche Antwort. Der Benutzer-Agent sollte eine davon wahlen.",
+  "301": "Die URL der angeforderten Ressource wurde dauerhaft geandert.",
+  "302": "Die URL der angeforderten Ressource wurde vorubergehend geandert.",
+  "304": "Gibt an, dass die Ressource seit der von den Anfrage-Headern angegebenen Version nicht geandert wurde.",
+  "307": "Der Server leitet den Benutzer-Agent zu einer anderen Ressource um, wie durch einen Location-Header angegeben.",
+  "308": "Die Ressource befindet sich jetzt dauerhaft an einer anderen URL, die durch den Location-Header angegeben ist.",
+  "400": "Der Server kann die Anfrage aufgrund eines offensichtlichen Client-Fehlers nicht oder wird sie nicht verarbeiten.",
+  "401": "Authentifizierung ist erforderlich und fehlgeschlagen oder wurde noch nicht bereitgestellt.",
+  "403": "Die Anfrage war gultig, aber der Server weigert sich zu handeln.",
+  "404": "Die angeforderte Ressource konnte nicht gefunden werden, konnte aber in Zukunft verfugbar sein.",
+  "405": "Eine Anfragemethode wird fur die angeforderte Ressource nicht unterstutzt.",
+  "406": "Die angeforderte Ressource kann nur Inhalte generieren, die nach den in der Anfrage gesendeten Accept-Headern nicht akzeptabel sind.",
+  "408": "Der Server hat beim Warten auf die Anfrage eine Zeituberschreitung erreicht.",
+  "409": "Die Anfrage konnte aufgrund eines Konflikts in der Anfrage nicht verarbeitet werden.",
+  "410": "Die angeforderte Ressource ist nicht mehr verfugbar und wird nicht mehr verfugbar sein.",
+  "411": "Die Anfrage hat die Lange ihres Inhalts nicht angegeben, was von der angeforderten Ressource erforderlich ist.",
+  "412": "Der Server erfullt eine der Vorbedingungen nicht, die der Anfragende an die Anfrage gestellt hat.",
+  "413": "Die Anfrage ist groser als der Server bereit oder in der Lage ist zu verarbeiten.",
+  "414": "Die bereitgestellte URI war zu lang, um vom Server verarbeitet zu werden.",
+  "415": "Die Anfrageentitat hat einen Medientyp, den der Server oder die Ressource nicht unterstutzt.",
+  "416": "Der Client hat nach einem Teil der Datei gefragt, aber der Server kann diesen Teil nicht liefern.",
+  "417": "Der Server kann die Anforderungen des Expect-Anfrage-Header-Felds nicht erfullen.",
+  "418": "Dieser Code wurde 1998 als einer der traditionellen IETF-Aprilscherze definiert.",
+  "422": "Die Anfrage war wohlgeformt, konnte aber aufgrund semantischer Fehler nicht befolgt werden.",
+  "425": "Gibt an, dass der Server nicht bereit ist, das Risiko einzugehen, eine Anfrage zu verarbeiten, die moglicherweise wiedergegeben wird.",
+  "426": "Der Client sollte zu einem anderen Protokoll wie TLS/1.0 wechseln.",
+  "428": "Der Ursprungsserver erfordert, dass die Anfrage bedingt ist.",
+  "429": "Der Benutzer hat in einem bestimmten Zeitraum zu viele Anfragen gesendet.",
+  "431": "Der Server ist nicht bereit, die Anfrage zu verarbeiten, da ihre Header-Felder zu gro sind.",
+  "451": "Der Server verweigert den Zugriff auf die Ressource als Folge einer rechtlichen Forderung.",
+  "500": "Eine generische Fehlermeldung, die ausgegeben wird, wenn eine unerwartete Bedingung aufgetreten ist.",
+  "501": "Der Server erkennt die Anfragemethode entweder nicht oder es fehlt ihm die Fahigkeit, die Anfrage zu erfullen.",
+  "502": "Der Server fungierte als Gateway oder Proxy und erhielt eine ungultige Antwort vom Upstream-Server.",
+  "503": "Der Server ist derzeit nicht verfugbar (weil er uberlastet ist oder zur Wartung ausgeschaltet ist).",
+  "504": "Der Server fungierte als Gateway oder Proxy und erhielt keine rechtzeitige Antwort vom Upstream-Server.",
+  "505": "Der Server unterstutzt die in der Anfrage verwendete HTTP-Protokollversion nicht.",
+  "506": "Die transparente Inhaltsverhandlung fur die Anfrage fuhrt zu einer zirkularen Referenz.",
+  "507": "Der Server kann die Darstellung nicht speichern, die zur Erfullung der Anfrage benotigt wird.",
+  "508": "Der Server hat eine Endlosschleife beim Verarbeiten der Anfrage erkannt.",
+  "510": "Weitere Erweiterungen der Anfrage sind erforderlich, damit der Server sie erfullen kann.",
+  "511": "Der Client muss sich authentifizieren, um Netzwerkzugriff zu erhalten."
+}
