+++ tools/http-status-code-lookup/descriptions/en.json
@@ -1,0 +1,52 @@
+{
+  "100": "The server has received the request headers and the client should proceed to send the request body.",
+  "101": "The requester has asked the server to switch protocols and the server has agreed.",
+  "102": "The server has received and is processing the request, but no response is available yet.",
+  "103": "Used to return some response headers before final HTTP message.",
+  "200": "The request has succeeded. The meaning of success depends on the HTTP method.",
+  "201": "The request has been fulfilled and resulted in a new resource being created.",
+  "202": "The request has been accepted for processing, but the processing has not been completed.",
+  "204": "The server successfully processed the request and is not returning any content.",
+  "206": "The server is delivering only part of the resource due to a range header sent by the client.",
+  "300": "The request has more than one possible response. The user agent should choose one of them.",
+  "301": "The URL of the requested resource has been changed permanently.",
+  "302": "The URL of the requested resource has been changed temporarily.",
+  "304": "Indicates that the resource has not been modified since the version specified by the request headers.",
+  "307": "The server is redirecting the user agent to a different resource, as indicated by a Location header.",
+  "308": "The resource is now permanently located at another URL, specified by the Location header.",
+  "400": "The server cannot or will not process the request due to an apparent client error.",
+  "401": "Authentication is required and has failed or has not yet been provided.",
+  "403": "The request was valid, but the server is refusing action.",
+  "404": "The requested resource could not be found but may be available in the future.",
+  "405": "A request method is not supported for the requested resource.",
+  "406": "The requested resource is capable of generating only content not acceptable according to the Accept headers sent in the request.",
+  "408": "The server timed out waiting for the request.",
+  "409": "The request could not be processed because of conflict in the request.",
+  "410": "The resource requested is no longer available and will not be available again.",
+  "411": "The request did not specify the length of its content, which is required by the requested resource.",
+  "412": "The server does not meet one of the preconditions that the requester put on the request.",
+  "413": "The request is larger than the server is willing or able to process.",
+  "414": "The URI provided was too long for the server to process.",
+  "415": "The request entity has a media type which the server or resource does not support.",
+  "416": "The client has asked for a portion of the file, but the server cannot supply that portion.",
+  "417": "The server cannot meet the requirements of the Expect request-header field.",
+  "418": "This code was defined in 1998 as one of the traditional IETF April Fools jokes.",
+  "422": "The request was well-formed but was unable to be followed due to semantic errors.",
+  "425": "Indicates that the server is unwilling to risk processing a request that might be replayed.",
+  "426": "The client should switch to a different protocol such as TLS/1.0.",
+  "428": "The origin server requires the request to be conditional.",
+  "429": "The user has sent too many requests in a given amount of time.",
+  "431": "The server is unwilling to process the request because its header fields are too large.",
+  "451": "The server is denying access to the resource as a consequence of a legal demand.",
+  "500": "A generic error message, given when an unexpected condition was encountered.",
+  "501": "The server either does not recognize the request method, or it lacks the ability to fulfill the request.",
+  "502": "The server was acting as a gateway or proxy and received an invalid response from the upstream server.",
+  "503": "The server is currently unavailable (because it is overloaded or down for maintenance).",
+  "504": "The server was acting as a gateway or proxy and did not receive a timely response from the upstream server.",
+  "505": "The server does not support the HTTP protocol version used in the request.",
+  "506": "Transparent content negotiation for the request results in a circular reference.",
+  "507": "The server is unable to store the representation needed to complete the request.",
+  "508": "The server detected an infinite loop while processing the request.",
+  "510": "Further extensions to the request are required for the server to fulfill it.",
+  "511": "The client needs to authenticate to gain network access."
+}
