+++ tools/http-status-code-lookup/descriptions/es.json
@@ -1,0 +1,52 @@
+{
+  "100": "El servidor ha recibido los encabezados de la solicitud y el cliente debe proceder a enviar el cuerpo de la solicitud.",
+  "101": "El solicitante ha pedido al servidor que cambie de protocolo y el servidor ha aceptado.",
+  "102": "El servidor ha recibido y esta procesando la solicitud, pero aun no hay respuesta disponible.",
+  "103": "Se utiliza para devolver algunos encabezados de respuesta antes del mensaje HTTP final.",
+  "200": "La solicitud ha tenido exito. El significado del exito depende del metodo HTTP.",
+  "201": "La solicitud se ha cumplido y ha resultado en la creacion de un nuevo recurso.",
+  "202": "La solicitud ha sido aceptada para su procesamiento, pero el procesamiento no se ha completado.",
+  "204": "El servidor proceso exitosamente la solicitud y no esta devolviendo ningun contenido.",
+  "206": "El servidor esta entregando solo una parte del recurso debido a un encabezado de rango enviado por el cliente.",
+  "300": "La solicitud tiene mas de una respuesta posible. El agente de usuario debe elegir una de ellas.",
+  "301": "La URL del recurso solicitado ha sido cambiada permanentemente.",
+  "302": "La URL del recurso solicitado ha sido cambiada temporalmente.",
+  "304": "Indica que el recurso no ha sido modificado desde la version especificada por los encabezados de la solicitud.",
+  "307": "El servidor esta redirigiendo al agente de usuario a un recurso diferente, como se indica en un encabezado Location.",
+  "308": "El recurso ahora esta permanentemente ubicado en otra URL, especificada por el encabezado Location.",
+  "400": "El servidor no puede o no procesara la solicitud debido a un error aparente del cliente.",
+  "401": "Se requiere autenticacion y ha fallado o aun no se ha proporcionado.",
+  "403": "La solicitud era valida, pero el servidor se niega a realizar la accion.",
+  "404": "No se pudo encontrar el recurso solicitado, pero puede estar disponible en el futuro.",
+  "405": "Un metodo de solicitud no es compatible con el recurso solicitado.",
+  "406": "El recurso solicitado solo es capaz de generar contenido no aceptable segun los encabezados Accept enviados en la solicitud.",
+  "408": "El servidor agoto el tiempo de espera de la solicitud.",
+  "409": "La solicitud no se pudo procesar debido a un conflicto en la solicitud.",
+  "410": "El recurso solicitado ya no esta disponible y no volvera a estar disponible.",
+  "411": "La solicitud no especifico la longitud de su contenido, que es requerida por el recurso solicitado.",
+  "412": "El servidor no cumple con una de las precondiciones que el solicitante establecio en la solicitud.",
+  "413": "La solicitud es mas grande de lo que el servidor esta dispuesto o puede procesar.",
+  "414": "El URI proporcionado era demasiado largo para que el servidor lo procesara.",
+  "415": "La entidad de solicitud tiene un tipo de medio que el servidor o recurso no soporta.",
+  "416": "El cliente ha pedido una porcion del archivo, pero el servidor no puede suministrar esa porcion.",
+  "417": "El servidor no puede cumplir con los requisitos del campo Expect del encabezado de solicitud.",
+  "418": "Este codigo fue definido en 1998 como una de las bromas tradicionales del Día de los Inocentes de IETF.",
+  "422": "La solicitud estaba bien formada pero no se pudo seguir debido a errores semanticos.",
+  "425": "Indica que el servidor no esta dispuesto a arriesgarse a procesar una solicitud que podria ser reproducida.",
+  "426": "El cliente debe cambiar a un protocolo diferente como TLS/1.0.",
+  "428": "El servidor de origen requiere que la solicitud sea condicional.",
+  "429": "El usuario ha enviado demasiadas solicitudes en un periodo de tiempo determinado.",
+  "431": "El servidor no esta dispuesto a procesar la solicitud porque sus campos de encabezado son demasiado grandes.",
+  "451": "El servidor esta denegando el acceso al recurso como consecuencia de una demanda legal.",
+  "500": "Un mensaje de error generico, dado cuando se encontro una condicion inesperada.",
+  "501": "El servidor no reconoce el metodo de solicitud o carece de la capacidad para cumplir la solicitud.",
+  "502": "El servidor estaba actuando como puerta de enlace o proxy y recibio una respuesta invalida del servidor ascendente.",
+  "503": "El servidor no esta disponible actualmente (porque esta sobrecargado o en mantenimiento).",
+  "504": "El servidor estaba actuando como puerta de enlace o proxy y no recibio una respuesta oportuna del servidor ascendente.",
+  "505": "El servidor no soporta la version del protocolo HTTP utilizada en la solicitud.",
+  "506": "La negociacion transparente de contenido para la solicitud resulta en una referencia circular.",
+  "507": "El servidor no puede almacenar la representacion necesaria para completar la solicitud.",
+  "508": "El servidor detecto un bucle infinito mientras procesaba la solicitud.",
+  "510": "Se requieren extensiones adicionales a la solicitud para que el servidor la cumpla.",
+  "511": "El cliente necesita autenticarse para obtener acceso a la red."
+}
