+++ tools/http-status-code-lookup/descriptions/fr.json
@@ -1,0 +1,52 @@
+{
+  "100": "Le serveur a recu les en-tetes de la requete et le client doit proceder a l'envoi du corps de la requete.",
+  "101": "Le demandeur a demande au serveur de changer de protocole et le serveur a accepte.",
+  "102": "Le serveur a recu et traite la requete, mais aucune reponse n'est encore disponible.",
+  "103": "Utilise pour renvoyer certains en-tetes de reponse avant le message HTTP final.",
+  "200": "La requete a reussi. La signification du succes depend de la methode HTTP.",
+  "201": "La requete a ete satisfaite et a abouti a la creation d'une nouvelle ressource.",
+  "202": "La requete a ete acceptee pour traitement, mais le traitement n'est pas termine.",
+  "204": "Le serveur a traite avec succes la requete et ne renvoie aucun contenu.",
+  "206": "Le serveur ne livre qu'une partie de la ressource en raison d'un en-tete de plage envoye par le client.",
+  "300": "La requete a plus d'une reponse possible. L'agent utilisateur doit en choisir une.",
+  "301": "L'URL de la ressource demandee a ete modifiee de maniere permanente.",
+  "302": "L'URL de la ressource demandee a ete modifiee temporairement.",
+  "304": "Indique que la ressource n'a pas ete modifiee depuis la version specifiee par les en-tetes de la requete.",
+  "307": "Le serveur redirige l'agent utilisateur vers une ressource differente, comme indique par un en-tete Location.",
+  "308": "La ressource est maintenant situee de maniere permanente a une autre URL, specifiee par l'en-tete Location.",
+  "400": "Le serveur ne peut pas ou ne traitera pas la requete en raison d'une erreur apparente du client.",
+  "401": "L'authentification est requise et a echoue ou n'a pas encore ete fournie.",
+  "403": "La requete etait valide, mais le serveur refuse d'agir.",
+  "404": "La ressource demandee n'a pas pu etre trouvee mais pourrait etre disponible a l'avenir.",
+  "405": "Une methode de requete n'est pas prise en charge pour la ressource demandee.",
+  "406": "La ressource demandee est capable de generer uniquement du contenu non acceptable selon les en-tetes Accept envoyes dans la requete.",
+  "408": "Le serveur a depasse le delai d'attente de la requete.",
+  "409": "La requete n'a pas pu etre traitee en raison d'un conflit dans la requete.",
+  "410": "La ressource demandee n'est plus disponible et ne sera plus disponible.",
+  "411": "La requete n'a pas specifie la longueur de son contenu, ce qui est requis par la ressource demandee.",
+  "412": "Le serveur ne repond pas a l'une des preconditions que le demandeur a mises sur la requete.",
+  "413": "La requete est plus grande que ce que le serveur est pret ou capable de traiter.",
+  "414": "L'URI fourni etait trop long pour que le serveur le traite.",
+  "415": "L'entite de requete a un type de media que le serveur ou la ressource ne prend pas en charge.",
+  "416": "Le client a demande une partie du fichier, mais le serveur ne peut pas fournir cette partie.",
+  "417": "Le serveur ne peut pas repondre aux exigences du champ Expect de l'en-tete de requete.",
+  "418": "Ce code a ete defini en 1998 comme l'une des plaisanteries traditionnelles du poisson d'avril de l'IETF.",
+  "422": "La requete etait bien formee mais n'a pas pu etre suivie en raison d'erreurs semantiques.",
+  "425": "Indique que le serveur n'est pas pret a risquer de traiter une requete qui pourrait etre rejouee.",
+  "426": "Le client doit passer a un protocole different tel que TLS/1.0.",
+  "428": "Le serveur d'origine exige que la requete soit conditionnelle.",
+  "429": "L'utilisateur a envoye trop de requetes dans un laps de temps donne.",
+  "431": "Le serveur n'est pas pret a traiter la requete car ses champs d'en-tete sont trop grands.",
+  "451": "Le serveur refuse l'acces a la ressource en consequence d'une demande legale.",
+  "500": "Un message d'erreur generique, donne lorsqu'une condition inattendue a ete rencontree.",
+  "501": "Le serveur ne reconnait pas la methode de requete ou manque de la capacite de satisfaire la requete.",
+  "502": "Le serveur agissait comme une passerelle ou un proxy et a recu une reponse invalide du serveur en amont.",
+  "503": "Le serveur n'est actuellement pas disponible (car il est surcharge ou en maintenance).",
+  "504": "Le serveur agissait comme une passerelle ou un proxy et n'a pas recu de reponse opportune du serveur en amont.",
+  "505": "Le serveur ne prend pas en charge la version du protocole HTTP utilisee dans la requete.",
+  "506": "La negociation transparente du contenu pour la requete aboutit a une reference circulaire.",
+  "507": "Le serveur est incapable de stocker la representation necessaire pour completer la requete.",
+  "508": "Le serveur a detecte une boucle infinie lors du traitement de la requete.",
+  "510": "Des extensions supplementaires a la requete sont requises pour que le serveur la satisfasse.",
+  "511": "Le client doit s'authentifier pour obtenir l'acces au reseau."
+}
