+++ tools/http-status-code-lookup/descriptions/he.json
@@ -1,0 +1,52 @@
+{
+  "100": "השרת קיבל את כותרות הבקשה והלקוח צריך להמשיך לשלוח את גוף הבקשה.",
+  "101": "המבקש ביקש מהשרת להחליף פרוטוקולים והשרת הסכים.",
+  "102": "השרת קיבל ומעבד את הבקשה, אך עדיין אין תגובה זמינה.",
+  "103": "משמש להחזרת כמה כותרות תגובה לפני הודעת HTTP הסופית.",
+  "200": "הבקשה הצליחה. משמעות ההצלחה תלויה בשיטת HTTP.",
+  "201": "הבקשה מולאה והביאה ליצירת משאב חדש.",
+  "202": "הבקשה התקבלה לעיבוד, אך העיבוד לא הושלם.",
+  "204": "השרת עיבד בהצלחה את הבקשה ואינו מחזיר תוכן כלשהו.",
+  "206": "השרת מספק רק חלק מהמשאב עקב כותרת טווח שנשלחה על ידי הלקוח.",
+  "300": "לבקשה יש יותר מתגובה אפשרית אחת. סוכן המשתמש צריך לבחור אחת מהן.",
+  "301": "כתובת ה-URL של המשאב המבוקש שונתה לצמיתות.",
+  "302": "כתובת ה-URL של המשאב המבוקש שונתה זמנית.",
+  "304": "מציין שהמשאב לא שונה מאז הגרסה שצוינה על ידי כותרות הבקשה.",
+  "307": "השרת מפנה את סוכן המשתמש למשאב אחר, כפי שמצוין על ידי כותרת Location.",
+  "308": "המשאב נמצא כעת לצמיתות בכתובת URL אחרת, שצוינה על ידי כותרת Location.",
+  "400": "השרת לא יכול או לא יעבד את הבקשה עקב שגיאת לקוח ברורה.",
+  "401": "נדרש אימות ונכשל או עדיין לא סופק.",
+  "403": "הבקשה הייתה תקפה, אך השרת מסרב לפעול.",
+  "404": "לא ניתן למצוא את המשאב המבוקש אך הוא עשוי להיות זמין בעתיד.",
+  "405": "שיטת בקשה אינה נתמכת עבור המשאב המבוקש.",
+  "406": "המשאב המבוקש מסוגל ליצור רק תוכן שאינו מקובל לפי כותרות Accept שנשלחו בבקשה.",
+  "408": "השרת הגיע לזמן קצוב בזמן שהמתין לבקשה.",
+  "409": "לא ניתן היה לעבד את הבקשה עקב קונפליקט בבקשה.",
+  "410": "המשאב המבוקש אינו זמין עוד ולא יהיה זמין שוב.",
+  "411": "הבקשה לא ציינה את אורך התוכן שלה, הנדרש על ידי המשאב המבוקש.",
+  "412": "השרת לא עומד באחד מתנאי הקדם שהמבקש הציב על הבקשה.",
+  "413": "הבקשה גדולה יותר ממה שהשרת מוכן או מסוגל לעבד.",
+  "414": "ה-URI שסופק היה ארוך מדי לשרת לעבד.",
+  "415": "ישות הבקשה כוללת סוג מדיה שהשרת או המשאב אינם תומכים בו.",
+  "416": "הלקוח ביקש חלק מהקובץ, אך השרת לא יכול לספק את החלק הזה.",
+  "417": "השרת לא יכול לעמוד בדרישות שדה כותרת הבקשה Expect.",
+  "418": "קוד זה הוגדר ב-1998 כאחד מבדיחות אחד באפריל המסורתיות של IETF.",
+  "422": "הבקשה הייתה מעוצבת היטב אך לא ניתן היה לעקוב אחריה עקב שגיאות סמנטיות.",
+  "425": "מציין שהשרת אינו מוכן לקחת סיכון בעיבוד בקשה שעשויה להיות משוחזרת.",
+  "426": "הלקוח צריך לעבור לפרוטוקול אחר כמו TLS/1.0.",
+  "428": "שרת המקור דורש שהבקשה תהיה מותנית.",
+  "429": "המשתמש שלח יותר מדי בקשות בתקופת זמן נתונה.",
+  "431": "השרת אינו מוכן לעבד את הבקשה מכיוון ששדות הכותרת שלו גדולים מדי.",
+  "451": "השרת מסרב לגישה למשאב כתוצאה מדרישה משפטית.",
+  "500": "הודעת שגיאה כללית, הניתנת כאשר נתקלו בתנאי בלתי צפוי.",
+  "501": "השרת אינו מזהה את שיטת הבקשה או חסר לו היכולת למלא את הבקשה.",
+  "502": "השרת פעל כשער או פרוקסי וקיבל תגובה לא תקפה מהשרת במעלה הזרם.",
+  "503": "השרת אינו זמין כעת (מכיוון שהוא עמוס מדי או כבוי לתחזוקה).",
+  "504": "השרת פעל כשער או פרוקסי ולא קיבל תגובה בזמן מהשרת במעלה הזרם.",
+  "505": "השרת אינו תומך בגרסת פרוטוקול HTTP המשמשת בבקשה.",
+  "506": "משא ומתן תוכן שקוף עבור הבקשה מביא להפנייה מעגלית.",
+  "507": "השרת אינו מסוגל לאחסן את הייצוג הנדרש להשלמת הבקשה.",
+  "508": "השרת זיהה לולאה אינסופית בעת עיבוד הבקשה.",
+  "510": "נדרשות הרחבות נוספות לבקשה כדי שהשרת יוכל למלא אותה.",
+  "511": "הלקוח צריך לאמת כדי לקבל גישה לרשת."
+}
