+++ tools/http-status-code-lookup/descriptions/hi.json
@@ -1,0 +1,52 @@
+{
+  "100": "सर्वर ने अनुरोध हेडर प्राप्त कर लिए हैं और क्लाइंट को अनुरोध बॉडी भेजना जारी रखना चाहिए।",
+  "101": "अनुरोधकर्ता ने सर्वर से प्रोटोकॉल बदलने के लिए कहा है और सर्वर सहमत हो गया है।",
+  "102": "सर्वर ने अनुरोध प्राप्त किया है और इसे संसाधित कर रहा है, लेकिन अभी तक कोई प्रतिक्रिया उपलब्ध नहीं है।",
+  "103": "अंतिम HTTP संदेश से पहले कुछ प्रतिक्रिया हेडर लौटाने के लिए उपयोग किया जाता है।",
+  "200": "अनुरोध सफल रहा। सफलता का अर्थ HTTP विधि पर निर्भर करता है।",
+  "201": "अनुरोध पूरा हो गया है और इसके परिणामस्वरूप एक नया संसाधन बनाया गया है।",
+  "202": "अनुरोध को संसाधन के लिए स्वीकार कर लिया गया है, लेकिन संसाधन पूरा नहीं हुआ है।",
+  "204": "सर्वर ने अनुरोध को सफलतापूर्वक संसाधित किया है और कोई सामग्री वापस नहीं कर रहा है।",
+  "206": "क्लाइंट द्वारा भेजे गए रेंज हेडर के कारण सर्वर संसाधन का केवल एक हिस्सा वितरित कर रहा है।",
+  "300": "अनुरोध में एक से अधिक संभावित प्रतिक्रियाएं हैं। उपयोगकर्ता एजेंट को उनमें से एक चुनना चाहिए।",
+  "301": "अनुरोधित संसाधन का URL स्थायी रूप से बदल दिया गया है।",
+  "302": "अनुरोधित संसाधन का URL अस्थायी रूप से बदल दिया गया है।",
+  "304": "इंगित करता है कि अनुरोध हेडर द्वारा निर्दिष्ट संस्करण के बाद से संसाधन संशोधित नहीं किया गया है।",
+  "307": "सर्वर उपयोगकर्ता एजेंट को एक अलग संसाधन पर पुनर्निर्देशित कर रहा है, जैसा कि Location हेडर द्वारा इंगित किया गया है।",
+  "308": "संसाधन अब स्थायी रूप से दूसरे URL पर स्थित है, Location हेडर द्वारा निर्दिष्ट।",
+  "400": "स्पष्ट क्लाइंट त्रुटि के कारण सर्वर अनुरोध को संसाधित नहीं कर सकता या नहीं करेगा।",
+  "401": "प्रमाणीकरण आवश्यक है और विफल हो गया है या अभी तक प्रदान नहीं किया गया है।",
+  "403": "अनुरोध वैध था, लेकिन सर्वर कार्रवाई से इनकार कर रहा है।",
+  "404": "अनुरोधित संसाधन नहीं मिला लेकिन भविष्य में उपलब्ध हो सकता है।",
+  "405": "अनुरोधित संसाधन के लिए अनुरोध विधि समर्थित नहीं है।",
+  "406": "अनुरोधित संसाधन केवल उस सामग्री को उत्पन्न करने में सक्षम है जो अनुरोध में भेजे गए Accept हेडर के अनुसार स्वीकार्य नहीं है।",
+  "408": "सर्वर ने अनुरोध की प्रतीक्षा करते समय समय सीमा समाप्त कर दी।",
+  "409": "अनुरोध में संघर्ष के कारण अनुरोध को संसाधित नहीं किया जा सका।",
+  "410": "अनुरोधित संसाधन अब उपलब्ध नहीं है और फिर से उपलब्ध नहीं होगा।",
+  "411": "अनुरोध ने अपनी सामग्री की लंबाई निर्दिष्ट नहीं की, जो अनुरोधित संसाधन द्वारा आवश्यक है।",
+  "412": "सर्वर उन पूर्व शर्तों में से एक को पूरा नहीं करता है जो अनुरोधकर्ता ने अनुरोध पर रखी थीं।",
+  "413": "अनुरोध सर्वर द्वारा संसाधित करने के लिए तैयार या सक्षम से अधिक बड़ा है।",
+  "414": "प्रदान किया गया URI सर्वर द्वारा संसाधित करने के लिए बहुत लंबा था।",
+  "415": "अनुरोध इकाई में एक मीडिया प्रकार है जिसे सर्वर या संसाधन समर्थन नहीं करता है।",
+  "416": "क्लाइंट ने फ़ाइल के एक हिस्से के लिए पूछा है, लेकिन सर्वर उस हिस्से को आपूर्ति नहीं कर सकता है।",
+  "417": "सर्वर अनुरोध हेडर फ़ील्ड Expect की आवश्यकताओं को पूरा नहीं कर सकता है।",
+  "418": "इस कोड को 1998 में पारंपरिक IETF अप्रैल फूल मजाक में से एक के रूप में परिभाषित किया गया था।",
+  "422": "अनुरोध अच्छी तरह से गठित था लेकिन अर्थ संबंधी त्रुटियों के कारण पालन नहीं किया जा सका।",
+  "425": "इंगित करता है कि सर्वर एक अनुरोध को संसाधित करने का जोखिम उठाने को तैयार नहीं है जिसे फिर से चलाया जा सकता है।",
+  "426": "क्लाइंट को TLS/1.0 जैसे अलग प्रोटोकॉल पर स्विच करना चाहिए।",
+  "428": "मूल सर्वर के लिए अनुरोध को सशर्त होना आवश्यक है।",
+  "429": "उपयोगकर्ता ने दिए गए समय में बहुत सारे अनुरोध भेजे हैं।",
+  "431": "सर्वर अनुरोध को संसाधित करने को तैयार नहीं है क्योंकि इसके हेडर फ़ील्ड बहुत बड़े हैं।",
+  "451": "सर्वर कानूनी मांग के परिणामस्वरूप संसाधन तक पहुंच से इनकार कर रहा है।",
+  "500": "एक सामान्य त्रुटि संदेश, जब एक अप्रत्याशित स्थिति का सामना किया गया था।",
+  "501": "सर्वर या तो अनुरोध विधि को पहचानता नहीं है, या अनुरोध को पूरा करने की क्षमता की कमी है।",
+  "502": "सर्वर गेटवे या प्रॉक्सी के रूप में कार्य कर रहा था और अपस्ट्रीम सर्वर से एक अमान्य प्रतिक्रिया प्राप्त की।",
+  "503": "सर्वर वर्तमान में उपलब्ध नहीं है (क्योंकि यह अधिक लोड है या रखरखाव के लिए बंद है)।",
+  "504": "सर्वर गेटवे या प्रॉक्सी के रूप में कार्य कर रहा था और अपस्ट्रीम सर्वर से समय पर प्रतिक्रिया प्राप्त नहीं की।",
+  "505": "सर्वर अनुरोध में उपयोग किए गए HTTP प्रोटोकॉल संस्करण का समर्थन नहीं करता है।",
+  "506": "अनुरोध के लिए पारदर्शी सामग्री वार्ता एक चक्रीय संदर्भ में परिणाम देती है।",
+  "507": "सर्वर अनुरोध को पूरा करने के लिए आवश्यक प्रतिनिधित्व को संग्रहीत करने में असमर्थ है।",
+  "508": "सर्वर ने अनुरोध को संसाधित करते समय एक अनंत लूप का पता लगाया।",
+  "510": "सर्वर को इसे पूरा करने के लिए अनुरोध के लिए आगे के विस्तार की आवश्यकता है।",
+  "511": "क्लाइंट को नेटवर्क एक्सेस प्राप्त करने के लिए प्रमाणित करने की आवश्यकता है।"
+}
