+++ tools/http-status-code-lookup/descriptions/id.json
@@ -1,0 +1,52 @@
+{
+  "100": "Server telah menerima header permintaan dan klien harus melanjutkan untuk mengirim body permintaan.",
+  "101": "Pemohon telah meminta server untuk beralih protokol dan server telah menyetujui.",
+  "102": "Server telah menerima dan sedang memproses permintaan, tetapi belum ada respons yang tersedia.",
+  "103": "Digunakan untuk mengembalikan beberapa header respons sebelum pesan HTTP akhir.",
+  "200": "Permintaan berhasil. Arti keberhasilan bergantung pada metode HTTP.",
+  "201": "Permintaan telah dipenuhi dan menghasilkan pembuatan sumber daya baru.",
+  "202": "Permintaan telah diterima untuk diproses, tetapi pemrosesan belum selesai.",
+  "204": "Server berhasil memproses permintaan dan tidak mengembalikan konten apa pun.",
+  "206": "Server hanya mengirimkan sebagian sumber daya karena header rentang yang dikirim oleh klien.",
+  "300": "Permintaan memiliki lebih dari satu respons yang mungkin. Agen pengguna harus memilih salah satunya.",
+  "301": "URL sumber daya yang diminta telah diubah secara permanen.",
+  "302": "URL sumber daya yang diminta telah diubah sementara.",
+  "304": "Menunjukkan bahwa sumber daya belum dimodifikasi sejak versi yang ditentukan oleh header permintaan.",
+  "307": "Server mengarahkan ulang agen pengguna ke sumber daya yang berbeda, seperti yang ditunjukkan oleh header Location.",
+  "308": "Sumber daya sekarang secara permanen berada di URL lain, yang ditentukan oleh header Location.",
+  "400": "Server tidak dapat atau tidak akan memproses permintaan karena kesalahan klien yang jelas.",
+  "401": "Autentikasi diperlukan dan telah gagal atau belum disediakan.",
+  "403": "Permintaan valid, tetapi server menolak untuk mengambil tindakan.",
+  "404": "Sumber daya yang diminta tidak dapat ditemukan tetapi mungkin tersedia di masa depan.",
+  "405": "Metode permintaan tidak didukung untuk sumber daya yang diminta.",
+  "406": "Sumber daya yang diminta hanya mampu menghasilkan konten yang tidak dapat diterima menurut header Accept yang dikirim dalam permintaan.",
+  "408": "Server mencapai batas waktu saat menunggu permintaan.",
+  "409": "Permintaan tidak dapat diproses karena konflik dalam permintaan.",
+  "410": "Sumber daya yang diminta tidak lagi tersedia dan tidak akan tersedia lagi.",
+  "411": "Permintaan tidak menentukan panjang kontennya, yang diperlukan oleh sumber daya yang diminta.",
+  "412": "Server tidak memenuhi salah satu prasyarat yang ditetapkan pemohon pada permintaan.",
+  "413": "Permintaan lebih besar daripada yang bersedia atau mampu diproses oleh server.",
+  "414": "URI yang diberikan terlalu panjang untuk diproses oleh server.",
+  "415": "Entitas permintaan memiliki jenis media yang tidak didukung oleh server atau sumber daya.",
+  "416": "Klien telah meminta sebagian file, tetapi server tidak dapat menyediakan bagian tersebut.",
+  "417": "Server tidak dapat memenuhi persyaratan bidang header permintaan Expect.",
+  "418": "Kode ini didefinisikan pada tahun 1998 sebagai salah satu lelucon tradisional April Mop IETF.",
+  "422": "Permintaan terbentuk dengan baik tetapi tidak dapat diikuti karena kesalahan semantik.",
+  "425": "Menunjukkan bahwa server tidak bersedia mengambil risiko memproses permintaan yang mungkin diputar ulang.",
+  "426": "Klien harus beralih ke protokol yang berbeda seperti TLS/1.0.",
+  "428": "Server asal memerlukan permintaan untuk bersyarat.",
+  "429": "Pengguna telah mengirim terlalu banyak permintaan dalam periode waktu tertentu.",
+  "431": "Server tidak bersedia memproses permintaan karena bidang headernya terlalu besar.",
+  "451": "Server menolak akses ke sumber daya sebagai konsekuensi dari permintaan hukum.",
+  "500": "Pesan kesalahan generik, diberikan ketika kondisi yang tidak terduga ditemui.",
+  "501": "Server tidak mengenali metode permintaan atau tidak memiliki kemampuan untuk memenuhi permintaan.",
+  "502": "Server bertindak sebagai gateway atau proxy dan menerima respons yang tidak valid dari server upstream.",
+  "503": "Server saat ini tidak tersedia (karena kelebihan beban atau sedang dalam pemeliharaan).",
+  "504": "Server bertindak sebagai gateway atau proxy dan tidak menerima respons tepat waktu dari server upstream.",
+  "505": "Server tidak mendukung versi protokol HTTP yang digunakan dalam permintaan.",
+  "506": "Negosiasi konten transparan untuk permintaan menghasilkan referensi melingkar.",
+  "507": "Server tidak dapat menyimpan representasi yang diperlukan untuk menyelesaikan permintaan.",
+  "508": "Server mendeteksi loop tak terbatas saat memproses permintaan.",
+  "510": "Ekstensi lebih lanjut ke permintaan diperlukan agar server dapat memenuhinya.",
+  "511": "Klien perlu mengautentikasi untuk mendapatkan akses jaringan."
+}
