+++ tools/http-status-code-lookup/descriptions/it.json
@@ -1,0 +1,52 @@
+{
+  "100": "Il server ha ricevuto le intestazioni della richiesta e il client deve procedere all'invio del corpo della richiesta.",
+  "101": "Il richiedente ha chiesto al server di cambiare protocollo e il server ha accettato.",
+  "102": "Il server ha ricevuto e sta elaborando la richiesta, ma non e ancora disponibile alcuna risposta.",
+  "103": "Utilizzato per restituire alcune intestazioni di risposta prima del messaggio HTTP finale.",
+  "200": "La richiesta ha avuto successo. Il significato del successo dipende dal metodo HTTP.",
+  "201": "La richiesta e stata soddisfatta e ha portato alla creazione di una nuova risorsa.",
+  "202": "La richiesta e stata accettata per l'elaborazione, ma l'elaborazione non e stata completata.",
+  "204": "Il server ha elaborato con successo la richiesta e non sta restituendo alcun contenuto.",
+  "206": "Il server sta consegnando solo una parte della risorsa a causa di un'intestazione di intervallo inviata dal client.",
+  "300": "La richiesta ha piu di una risposta possibile. L'agente utente dovrebbe sceglierne una.",
+  "301": "L'URL della risorsa richiesta e stata modificata permanentemente.",
+  "302": "L'URL della risorsa richiesta e stata modificata temporaneamente.",
+  "304": "Indica che la risorsa non e stata modificata dalla versione specificata dalle intestazioni della richiesta.",
+  "307": "Il server sta reindirizzando l'agente utente a una risorsa diversa, come indicato da un'intestazione Location.",
+  "308": "La risorsa e ora permanentemente situata in un altro URL, specificato dall'intestazione Location.",
+  "400": "Il server non puo o non elaborera la richiesta a causa di un evidente errore del client.",
+  "401": "L'autenticazione e richiesta e e fallita o non e ancora stata fornita.",
+  "403": "La richiesta era valida, ma il server rifiuta di agire.",
+  "404": "La risorsa richiesta non e stata trovata ma potrebbe essere disponibile in futuro.",
+  "405": "Un metodo di richiesta non e supportato per la risorsa richiesta.",
+  "406": "La risorsa richiesta e in grado di generare solo contenuti non accettabili secondo le intestazioni Accept inviate nella richiesta.",
+  "408": "Il server ha raggiunto il timeout in attesa della richiesta.",
+  "409": "La richiesta non poteva essere elaborata a causa di un conflitto nella richiesta.",
+  "410": "La risorsa richiesta non e piu disponibile e non sara piu disponibile.",
+  "411": "La richiesta non ha specificato la lunghezza del suo contenuto, che e richiesta dalla risorsa richiesta.",
+  "412": "Il server non soddisfa una delle precondizioni che il richiedente ha posto sulla richiesta.",
+  "413": "La richiesta e piu grande di quanto il server sia disposto o in grado di elaborare.",
+  "414": "L'URI fornito era troppo lungo per essere elaborato dal server.",
+  "415": "L'entita della richiesta ha un tipo di supporto che il server o la risorsa non supporta.",
+  "416": "Il client ha chiesto una porzione del file, ma il server non puo fornire quella porzione.",
+  "417": "Il server non puo soddisfare i requisiti del campo Expect dell'intestazione della richiesta.",
+  "418": "Questo codice e stato definito nel 1998 come uno degli scherzi tradizionali del pesce d'aprile dell'IETF.",
+  "422": "La richiesta era ben formata ma non poteva essere seguita a causa di errori semantici.",
+  "425": "Indica che il server non e disposto a rischiare di elaborare una richiesta che potrebbe essere riprodotta.",
+  "426": "Il client dovrebbe passare a un protocollo diverso come TLS/1.0.",
+  "428": "Il server di origine richiede che la richiesta sia condizionale.",
+  "429": "L'utente ha inviato troppe richieste in un determinato periodo di tempo.",
+  "431": "Il server non e disposto a elaborare la richiesta perche i suoi campi di intestazione sono troppo grandi.",
+  "451": "Il server sta negando l'accesso alla risorsa come conseguenza di una richiesta legale.",
+  "500": "Un messaggio di errore generico, dato quando e stata incontrata una condizione inaspettata.",
+  "501": "Il server non riconosce il metodo di richiesta o manca della capacita di soddisfare la richiesta.",
+  "502": "Il server stava agendo come gateway o proxy e ha ricevuto una risposta non valida dal server upstream.",
+  "503": "Il server non e attualmente disponibile (perche e sovraccarico o in manutenzione).",
+  "504": "Il server stava agendo come gateway o proxy e non ha ricevuto una risposta tempestiva dal server upstream.",
+  "505": "Il server non supporta la versione del protocollo HTTP utilizzata nella richiesta.",
+  "506": "La negoziazione trasparente del contenuto per la richiesta risulta in un riferimento circolare.",
+  "507": "Il server e incapace di memorizzare la rappresentazione necessaria per completare la richiesta.",
+  "508": "Il server ha rilevato un ciclo infinito durante l'elaborazione della richiesta.",
+  "510": "Sono richieste ulteriori estensioni alla richiesta affinche il server la soddisfi.",
+  "511": "Il client deve autenticarsi per ottenere l'accesso alla rete."
+}
