+++ tools/http-status-code-lookup/descriptions/ja.json
@@ -1,0 +1,52 @@
+{
+  "100": "サーバーはリクエストヘッダーを受信しました。クライアントはリクエストボディの送信を続行する必要があります。",
+  "101": "リクエスターはサーバーにプロトコルの切り替えを要求し、サーバーは同意しました。",
+  "102": "サーバーはリクエストを受信して処理中ですが、まだ応答は利用できません。",
+  "103": "最終的なHTTPメッセージの前に、いくつかの応答ヘッダーを返すために使用されます。",
+  "200": "リクエストは成功しました。成功の意味はHTTPメソッドによって異なります。",
+  "201": "リクエストは完了し、新しいリソースの作成につながりました。",
+  "202": "リクエストは処理のために受け入れられましたが、処理はまだ完了していません。",
+  "204": "サーバーはリクエストを正常に処理しましたが、コンテンツを返していません。",
+  "206": "クライアントが送信した範囲ヘッダーにより、サーバーはリソースの一部のみを配信しています。",
+  "300": "リクエストには複数の可能な応答があります。ユーザーエージェントはそのうちの1つを選択する必要があります。",
+  "301": "要求されたリソースのURLが永続的に変更されました。",
+  "302": "要求されたリソースのURLが一時的に変更されました。",
+  "304": "リクエストヘッダーで指定されたバージョン以降、リソースが変更されていないことを示します。",
+  "307": "サーバーは、Locationヘッダーで示されているように、ユーザーエージェントを別のリソースにリダイレクトしています。",
+  "308": "リソースは現在、Locationヘッダーで指定された別のURLに永続的に配置されています。",
+  "400": "明らかなクライアントエラーのため、サーバーはリクエストを処理できません。",
+  "401": "認証が必要ですが、失敗したか、まだ提供されていません。",
+  "403": "リクエストは有効でしたが、サーバーはアクションを拒否しています。",
+  "404": "要求されたリソースが見つかりませんでしたが、将来利用可能になる可能性があります。",
+  "405": "要求されたリソースでは、リクエストメソッドがサポートされていません。",
+  "406": "要求されたリソースは、リクエストで送信されたAcceptヘッダーに従って受け入れられないコンテンツのみを生成できます。",
+  "408": "サーバーはリクエストの待機中にタイムアウトしました。",
+  "409": "リクエスト内の競合により、リクエストを処理できませんでした。",
+  "410": "要求されたリソースは利用できなくなり、今後も利用できません。",
+  "411": "リクエストはコンテンツの長さを指定していませんでしたが、これは要求されたリソースで必要です。",
+  "412": "サーバーは、リクエスターがリクエストに設定した前提条件の1つを満たしていません。",
+  "413": "リクエストは、サーバーが処理する意思があるか、処理できるサイズよりも大きいです。",
+  "414": "提供されたURIが長すぎて、サーバーが処理できませんでした。",
+  "415": "リクエストエンティティには、サーバーまたはリソースがサポートしていないメディアタイプがあります。",
+  "416": "クライアントはファイルの一部を要求しましたが、サーバーはその部分を提供できません。",
+  "417": "サーバーは、Expectリクエストヘッダーフィールドの要件を満たすことができません。",
+  "418": "このコードは1998年に、伝統的なIETFエイプリルフールのジョークの1つとして定義されました。",
+  "422": "リクエストは適切に形成されていましたが、セマンティックエラーにより従うことができませんでした。",
+  "425": "サーバーは、再生される可能性のあるリクエストを処理するリスクを負うことを望んでいないことを示します。",
+  "426": "クライアントは、TLS/1.0などの別のプロトコルに切り替える必要があります。",
+  "428": "オリジンサーバーは、リクエストが条件付きであることを要求します。",
+  "429": "ユーザーは、指定された時間内にリクエストを送信しすぎました。",
+  "431": "サーバーは、ヘッダーフィールドが大きすぎるため、リクエストを処理することを望んでいません。",
+  "451": "サーバーは、法的要求の結果としてリソースへのアクセスを拒否しています。",
+  "500": "予期しない条件が発生したときに与えられる一般的なエラーメッセージ。",
+  "501": "サーバーはリクエストメソッドを認識しないか、リクエストを満たす能力がありません。",
+  "502": "サーバーはゲートウェイまたはプロキシとして機能し、アップストリームサーバーから無効な応答を受信しました。",
+  "503": "サーバーは現在利用できません（過負荷またはメンテナンス中のため）。",
+  "504": "サーバーはゲートウェイまたはプロキシとして機能していましたが、アップストリームサーバーからタイムリーな応答を受信しませんでした。",
+  "505": "サーバーは、リクエストで使用されているHTTPプロトコルバージョンをサポートしていません。",
+  "506": "リクエストの透過的なコンテンツネゴシエーションは、循環参照になります。",
+  "507": "サーバーは、リクエストを完了するために必要な表現を保存できません。",
+  "508": "サーバーは、リクエストの処理中に無限ループを検出しました。",
+  "510": "サーバーがリクエストを満たすには、リクエストへのさらなる拡張が必要です。",
+  "511": "クライアントはネットワークアクセスを取得するために認証する必要があります。"
+}
