+++ tools/http-status-code-lookup/descriptions/ko.json
@@ -1,0 +1,52 @@
+{
+  "100": "서버가 요청 헤더를 받았으며 클라이언트는 요청 본문 전송을 계속해야 합니다.",
+  "101": "요청자가 서버에 프로토콜 전환을 요청했고 서버가 동의했습니다.",
+  "102": "서버가 요청을 받아 처리 중이지만 아직 응답을 사용할 수 없습니다.",
+  "103": "최종 HTTP 메시지 전에 일부 응답 헤더를 반환하는 데 사용됩니다.",
+  "200": "요청이 성공했습니다. 성공의 의미는 HTTP 메서드에 따라 다릅니다.",
+  "201": "요청이 완료되어 새 리소스가 생성되었습니다.",
+  "202": "요청이 처리되도록 수락되었지만 처리가 완료되지 않았습니다.",
+  "204": "서버가 요청을 성공적으로 처리했으며 콘텐츠를 반환하지 않습니다.",
+  "206": "클라이언트가 보낸 범위 헤더로 인해 서버가 리소스의 일부만 전달하고 있습니다.",
+  "300": "요청에 여러 가지 가능한 응답이 있습니다. 사용자 에이전트는 그 중 하나를 선택해야 합니다.",
+  "301": "요청된 리소스의 URL이 영구적으로 변경되었습니다.",
+  "302": "요청된 리소스의 URL이 일시적으로 변경되었습니다.",
+  "304": "요청 헤더로 지정된 버전 이후 리소스가 수정되지 않았음을 나타냅니다.",
+  "307": "서버가 Location 헤더로 표시된 대로 사용자 에이전트를 다른 리소스로 리디렉션하고 있습니다.",
+  "308": "리소스가 이제 Location 헤더로 지정된 다른 URL에 영구적으로 위치합니다.",
+  "400": "명백한 클라이언트 오류로 인해 서버가 요청을 처리할 수 없거나 처리하지 않습니다.",
+  "401": "인증이 필요하며 실패했거나 아직 제공되지 않았습니다.",
+  "403": "요청이 유효했지만 서버가 작업을 거부하고 있습니다.",
+  "404": "요청된 리소스를 찾을 수 없지만 나중에 사용할 수 있을 수 있습니다.",
+  "405": "요청된 리소스에 대해 요청 메서드가 지원되지 않습니다.",
+  "406": "요청된 리소스는 요청에서 보낸 Accept 헤더에 따라 허용되지 않는 콘텐츠만 생성할 수 있습니다.",
+  "408": "서버가 요청 대기 중 시간 초과되었습니다.",
+  "409": "요청의 충돌로 인해 요청을 처리할 수 없었습니다.",
+  "410": "요청된 리소스가 더 이상 사용할 수 없으며 다시 사용할 수 없습니다.",
+  "411": "요청이 콘텐츠 길이를 지정하지 않았으며, 이는 요청된 리소스에 필요합니다.",
+  "412": "서버가 요청자가 요청에 설정한 전제 조건 중 하나를 충족하지 않습니다.",
+  "413": "요청이 서버가 처리하려고 하거나 처리할 수 있는 것보다 큽니다.",
+  "414": "제공된 URI가 너무 길어 서버가 처리할 수 없었습니다.",
+  "415": "요청 엔터티에 서버 또는 리소스가 지원하지 않는 미디어 유형이 있습니다.",
+  "416": "클라이언트가 파일의 일부를 요청했지만 서버가 해당 부분을 제공할 수 없습니다.",
+  "417": "서버가 Expect 요청 헤더 필드의 요구 사항을 충족할 수 없습니다.",
+  "418": "이 코드는 1998년 전통적인 IETF 만우절 농담 중 하나로 정의되었습니다.",
+  "422": "요청이 잘 형성되었지만 의미론적 오류로 인해 따를 수 없었습니다.",
+  "425": "서버가 재생될 수 있는 요청을 처리하는 위험을 감수하지 않으려 함을 나타냅니다.",
+  "426": "클라이언트는 TLS/1.0과 같은 다른 프로토콜로 전환해야 합니다.",
+  "428": "원본 서버는 요청이 조건부여야 합니다.",
+  "429": "사용자가 지정된 시간 내에 너무 많은 요청을 보냈습니다.",
+  "431": "서버가 헤더 필드가 너무 커서 요청을 처리하지 않으려 합니다.",
+  "451": "서버가 법적 요구의 결과로 리소스에 대한 액세스를 거부하고 있습니다.",
+  "500": "예기치 않은 조건이 발생했을 때 제공되는 일반 오류 메시지입니다.",
+  "501": "서버가 요청 메서드를 인식하지 못하거나 요청을 수행할 능력이 없습니다.",
+  "502": "서버가 게이트웨이 또는 프록시로 작동했으며 업스트림 서버로부터 잘못된 응답을 받았습니다.",
+  "503": "서버가 현재 사용할 수 없습니다(과부하 또는 유지 관리 중이기 때문).",
+  "504": "서버가 게이트웨이 또는 프록시로 작동했지만 업스트림 서버로부터 적시에 응답을 받지 못했습니다.",
+  "505": "서버가 요청에 사용된 HTTP 프로토콜 버전을 지원하지 않습니다.",
+  "506": "요청에 대한 투명한 콘텐츠 협상이 순환 참조로 이어집니다.",
+  "507": "서버가 요청을 완료하는 데 필요한 표현을 저장할 수 없습니다.",
+  "508": "서버가 요청 처리 중 무한 루프를 감지했습니다.",
+  "510": "서버가 요청을 수행하려면 요청에 대한 추가 확장이 필요합니다.",
+  "511": "클라이언트가 네트워크 액세스를 얻으려면 인증해야 합니다."
+}
