+++ tools/http-status-code-lookup/descriptions/ms.json
@@ -1,0 +1,52 @@
+{
+  "100": "Pelayan telah menerima pengepala permintaan dan klien harus meneruskan untuk menghantar badan permintaan.",
+  "101": "Pemohon telah meminta pelayan menukar protokol dan pelayan telah bersetuju.",
+  "102": "Pelayan telah menerima dan sedang memproses permintaan, tetapi tiada respons tersedia lagi.",
+  "103": "Digunakan untuk mengembalikan beberapa pengepala respons sebelum mesej HTTP akhir.",
+  "200": "Permintaan berjaya. Maksud kejayaan bergantung pada kaedah HTTP.",
+  "201": "Permintaan telah dipenuhi dan mengakibatkan penciptaan sumber baharu.",
+  "202": "Permintaan telah diterima untuk diproses, tetapi pemprosesan belum selesai.",
+  "204": "Pelayan berjaya memproses permintaan dan tidak mengembalikan sebarang kandungan.",
+  "206": "Pelayan hanya menyampaikan sebahagian sumber kerana pengepala julat yang dihantar oleh klien.",
+  "300": "Permintaan mempunyai lebih daripada satu respons yang mungkin. Ejen pengguna harus memilih salah satu daripadanya.",
+  "301": "URL sumber yang diminta telah diubah secara kekal.",
+  "302": "URL sumber yang diminta telah diubah sementara.",
+  "304": "Menunjukkan bahawa sumber tidak diubah sejak versi yang ditentukan oleh pengepala permintaan.",
+  "307": "Pelayan mengalihkan ejen pengguna ke sumber yang berbeza, seperti yang ditunjukkan oleh pengepala Location.",
+  "308": "Sumber kini terletak secara kekal di URL lain, yang ditentukan oleh pengepala Location.",
+  "400": "Pelayan tidak boleh atau tidak akan memproses permintaan kerana ralat klien yang jelas.",
+  "401": "Pengesahan diperlukan dan telah gagal atau belum disediakan.",
+  "403": "Permintaan adalah sah, tetapi pelayan menolak untuk bertindak.",
+  "404": "Sumber yang diminta tidak dapat ditemui tetapi mungkin tersedia pada masa hadapan.",
+  "405": "Kaedah permintaan tidak disokong untuk sumber yang diminta.",
+  "406": "Sumber yang diminta hanya mampu menjana kandungan yang tidak boleh diterima mengikut pengepala Accept yang dihantar dalam permintaan.",
+  "408": "Pelayan mencapai masa tamat semasa menunggu permintaan.",
+  "409": "Permintaan tidak dapat diproses kerana konflik dalam permintaan.",
+  "410": "Sumber yang diminta tidak lagi tersedia dan tidak akan tersedia lagi.",
+  "411": "Permintaan tidak menentukan panjang kandungannya, yang diperlukan oleh sumber yang diminta.",
+  "412": "Pelayan tidak memenuhi salah satu prasyarat yang ditetapkan pemohon pada permintaan.",
+  "413": "Permintaan lebih besar daripada yang pelayan sanggup atau mampu memproses.",
+  "414": "URI yang diberikan terlalu panjang untuk pelayan memproses.",
+  "415": "Entiti permintaan mempunyai jenis media yang tidak disokong oleh pelayan atau sumber.",
+  "416": "Klien telah meminta sebahagian fail, tetapi pelayan tidak dapat membekalkan bahagian tersebut.",
+  "417": "Pelayan tidak dapat memenuhi keperluan medan pengepala permintaan Expect.",
+  "418": "Kod ini ditakrifkan pada tahun 1998 sebagai salah satu jenaka April Fool tradisional IETF.",
+  "422": "Permintaan dibentuk dengan baik tetapi tidak dapat diikuti kerana ralat semantik.",
+  "425": "Menunjukkan bahawa pelayan tidak sanggup mengambil risiko memproses permintaan yang mungkin dimainkan semula.",
+  "426": "Klien harus beralih kepada protokol yang berbeza seperti TLS/1.0.",
+  "428": "Pelayan asal memerlukan permintaan untuk bersyarat.",
+  "429": "Pengguna telah menghantar terlalu banyak permintaan dalam tempoh masa tertentu.",
+  "431": "Pelayan tidak sanggup memproses permintaan kerana medan pengepalanya terlalu besar.",
+  "451": "Pelayan menolak akses kepada sumber sebagai akibat daripada permintaan undang-undang.",
+  "500": "Mesej ralat generik, diberikan apabila keadaan yang tidak dijangka ditemui.",
+  "501": "Pelayan tidak mengenali kaedah permintaan atau tidak mempunyai keupayaan untuk memenuhi permintaan.",
+  "502": "Pelayan bertindak sebagai pintu masuk atau proksi dan menerima respons yang tidak sah daripada pelayan hulu.",
+  "503": "Pelayan tidak tersedia pada masa ini (kerana terlebih muatan atau ditutup untuk penyelenggaraan).",
+  "504": "Pelayan bertindak sebagai pintu masuk atau proksi dan tidak menerima respons tepat pada masanya daripada pelayan hulu.",
+  "505": "Pelayan tidak menyokong versi protokol HTTP yang digunakan dalam permintaan.",
+  "506": "Perundingan kandungan telus untuk permintaan mengakibatkan rujukan bulat.",
+  "507": "Pelayan tidak dapat menyimpan perwakilan yang diperlukan untuk melengkapkan permintaan.",
+  "508": "Pelayan mengesan gelung tak terhingga semasa memproses permintaan.",
+  "510": "Sambungan lanjut kepada permintaan diperlukan untuk pelayan memenuhinya.",
+  "511": "Klien perlu mengesahkan untuk mendapatkan akses rangkaian."
+}
