+++ tools/http-status-code-lookup/descriptions/nl.json
@@ -1,0 +1,52 @@
+{
+  "100": "De server heeft de verzoekheaders ontvangen en de client moet doorgaan met het verzenden van de verzoekbody.",
+  "101": "De aanvrager heeft de server gevraagd om protocollen te wisselen en de server heeft ingestemd.",
+  "102": "De server heeft het verzoek ontvangen en verwerkt het, maar er is nog geen antwoord beschikbaar.",
+  "103": "Gebruikt om enkele antwoordheaders terug te geven vóór het definitieve HTTP-bericht.",
+  "200": "Het verzoek is geslaagd. De betekenis van succes hangt af van de HTTP-methode.",
+  "201": "Het verzoek is vervuld en heeft geleid tot het creëren van een nieuwe bron.",
+  "202": "Het verzoek is geaccepteerd voor verwerking, maar de verwerking is nog niet voltooid.",
+  "204": "De server heeft het verzoek succesvol verwerkt en retourneert geen inhoud.",
+  "206": "De server levert slechts een deel van de bron vanwege een bereikheader die door de client is verzonden.",
+  "300": "Het verzoek heeft meer dan één mogelijk antwoord. De user agent moet er één kiezen.",
+  "301": "De URL van de gevraagde bron is permanent gewijzigd.",
+  "302": "De URL van de gevraagde bron is tijdelijk gewijzigd.",
+  "304": "Geeft aan dat de bron niet is gewijzigd sinds de versie die is opgegeven door de verzoekheaders.",
+  "307": "De server leidt de user agent om naar een andere bron, zoals aangegeven door een Location-header.",
+  "308": "De bron bevindt zich nu permanent op een andere URL, opgegeven door de Location-header.",
+  "400": "De server kan het verzoek niet of zal het niet verwerken vanwege een duidelijk clientfout.",
+  "401": "Authenticatie is vereist en is mislukt of nog niet verstrekt.",
+  "403": "Het verzoek was geldig, maar de server weigert actie te ondernemen.",
+  "404": "De gevraagde bron kon niet worden gevonden maar kan in de toekomst beschikbaar zijn.",
+  "405": "Een verzoekmethode wordt niet ondersteund voor de gevraagde bron.",
+  "406": "De gevraagde bron kan alleen inhoud genereren die niet acceptabel is volgens de Accept-headers die in het verzoek zijn verzonden.",
+  "408": "De server heeft een time-out bereikt tijdens het wachten op het verzoek.",
+  "409": "Het verzoek kon niet worden verwerkt vanwege een conflict in het verzoek.",
+  "410": "De gevraagde bron is niet meer beschikbaar en zal niet meer beschikbaar zijn.",
+  "411": "Het verzoek heeft de lengte van de inhoud niet opgegeven, wat vereist is door de gevraagde bron.",
+  "412": "De server voldoet niet aan een van de voorwaarden die de aanvrager aan het verzoek heeft gesteld.",
+  "413": "Het verzoek is groter dan de server bereid is of kan verwerken.",
+  "414": "De verstrekte URI was te lang voor de server om te verwerken.",
+  "415": "De verzoekentiteit heeft een mediatype dat de server of bron niet ondersteunt.",
+  "416": "De client heeft om een deel van het bestand gevraagd, maar de server kan dat deel niet leveren.",
+  "417": "De server kan niet voldoen aan de vereisten van het Expect-verzoekheaderveld.",
+  "418": "Deze code werd in 1998 gedefinieerd als een van de traditionele IETF 1 april grappen.",
+  "422": "Het verzoek was goed gevormd maar kon niet worden gevolgd vanwege semantische fouten.",
+  "425": "Geeft aan dat de server niet bereid is het risico te nemen om een verzoek te verwerken dat mogelijk opnieuw wordt afgespeeld.",
+  "426": "De client moet overschakelen naar een ander protocol zoals TLS/1.0.",
+  "428": "De oorspronkelijke server vereist dat het verzoek voorwaardelijk is.",
+  "429": "De gebruiker heeft te veel verzoeken verzonden in een bepaalde tijdsperiode.",
+  "431": "De server is niet bereid het verzoek te verwerken omdat de headervelden te groot zijn.",
+  "451": "De server weigert toegang tot de bron als gevolg van een juridische eis.",
+  "500": "Een generiek foutbericht, gegeven wanneer een onverwachte voorwaarde werd aangetroffen.",
+  "501": "De server herkent de verzoekmethode niet of mist het vermogen om het verzoek te vervullen.",
+  "502": "De server fungeerde als gateway of proxy en ontving een ongeldig antwoord van de upstream-server.",
+  "503": "De server is momenteel niet beschikbaar (omdat deze overbelast is of voor onderhoud uitgeschakeld).",
+  "504": "De server fungeerde als gateway of proxy en ontving geen tijdig antwoord van de upstream-server.",
+  "505": "De server ondersteunt de HTTP-protocolversie die in het verzoek wordt gebruikt niet.",
+  "506": "Transparante inhoudsonderhandeling voor het verzoek resulteert in een circulaire referentie.",
+  "507": "De server kan de representatie die nodig is om het verzoek te voltooien niet opslaan.",
+  "508": "De server heeft een oneindige lus gedetecteerd tijdens het verwerken van het verzoek.",
+  "510": "Verdere uitbreidingen van het verzoek zijn vereist voor de server om het te vervullen.",
+  "511": "De client moet zich authenticeren om netwerktoegang te krijgen."
+}
