+++ tools/http-status-code-lookup/descriptions/no.json
@@ -1,0 +1,52 @@
+{
+  "100": "Serveren har mottatt foresporselshodene og klienten bor fortsette med å sende foresporselskroppen.",
+  "101": "Forespøreren har bedt serveren om å bytte protokoller og serveren har gått med på det.",
+  "102": "Serveren har mottatt og behandler forespørselen, men ingen respons er tilgjengelig ennå.",
+  "103": "Brukes til å returnere noen responshoder før den endelige HTTP-meldingen.",
+  "200": "Forespørselen var vellykket. Betydningen av suksess avhenger av HTTP-metoden.",
+  "201": "Forespørselen har blitt oppfylt og resulterte i at en ny ressurs ble opprettet.",
+  "202": "Forespørselen har blitt akseptert for behandling, men behandlingen er ikke fullført.",
+  "204": "Serveren behandlet forespørselen vellykket og returnerer ikke noe innhold.",
+  "206": "Serveren leverer bare en del av ressursen på grunn av et områdehode sendt av klienten.",
+  "300": "Forespørselen har mer enn ett mulig svar. Brukeragenten bør velge ett av dem.",
+  "301": "URL-en til den forespurte ressursen har blitt endret permanent.",
+  "302": "URL-en til den forespurte ressursen har blitt endret midlertidig.",
+  "304": "Angir at ressursen ikke har blitt endret siden versjonen spesifisert av foresporselshodene.",
+  "307": "Serveren omdirigerer brukeragenten til en annen ressurs, som angitt av et Location-hode.",
+  "308": "Ressursen er nå permanent lokalisert på en annen URL, spesifisert av Location-hodet.",
+  "400": "Serveren kan ikke eller vil ikke behandle forespørselen på grunn av en åpenbar klientfeil.",
+  "401": "Autentisering er påkrevd og har mislyktes eller har ikke blitt gitt ennå.",
+  "403": "Forespørselen var gyldig, men serveren nekter å handle.",
+  "404": "Den forespurte ressursen kunne ikke bli funnet, men kan være tilgjengelig i fremtiden.",
+  "405": "En foresporselsmetode støttes ikke for den forespurte ressursen.",
+  "406": "Den forespurte ressursen er bare i stand til å generere innhold som ikke er akseptabelt i henhold til Accept-hodene sendt i forespørselen.",
+  "408": "Serveren nådde tidsavbrudd mens den ventet på forespørselen.",
+  "409": "Forespørselen kunne ikke behandles på grunn av en konflikt i forespørselen.",
+  "410": "Den forespurte ressursen er ikke lenger tilgjengelig og vil ikke være tilgjengelig igjen.",
+  "411": "Forespørselen spesifiserte ikke lengden på innholdet, som kreves av den forespurte ressursen.",
+  "412": "Serveren oppfyller ikke en av forutsetningene som forespøreren satte på forespørselen.",
+  "413": "Forespørselen er større enn serveren er villig til eller i stand til å behandle.",
+  "414": "Den oppgitte URI-en var for lang for serveren å behandle.",
+  "415": "Foresporselsentiteten har en medietype som serveren eller ressursen ikke støtter.",
+  "416": "Klienten har bedt om en del av filen, men serveren kan ikke levere den delen.",
+  "417": "Serveren kan ikke oppfylle kravene til Expect-foresporselshodefeltet.",
+  "418": "Denne koden ble definert i 1998 som en av de tradisjonelle IETF-aprilspøkene.",
+  "422": "Forespørselen var velformet, men kunne ikke følges på grunn av semantiske feil.",
+  "425": "Angir at serveren ikke er villig til å risikere å behandle en forespørsel som kan bli spilt av igjen.",
+  "426": "Klienten bør bytte til en annen protokoll som TLS/1.0.",
+  "428": "Opprinnelsesserveren krever at forespørselen skal være betinget.",
+  "429": "Brukeren har sendt for mange forespørsler i en gitt tidsperiode.",
+  "431": "Serveren er ikke villig til å behandle forespørselen fordi hodefeltene er for store.",
+  "451": "Serveren nekter tilgang til ressursen som en konsekvens av en juridisk forespørsel.",
+  "500": "En generisk feilmelding, gitt når en uventet tilstand ble påtruffet.",
+  "501": "Serveren gjenkjenner enten ikke foresporselsmetoden eller mangler evnen til å oppfylle forespørselen.",
+  "502": "Serveren fungerte som en gateway eller proxy og mottok et ugyldig svar fra upstream-serveren.",
+  "503": "Serveren er for øyeblikket ikke tilgjengelig (fordi den er overbelastet eller nede for vedlikehold).",
+  "504": "Serveren fungerte som en gateway eller proxy og mottok ikke et rettidig svar fra upstream-serveren.",
+  "505": "Serveren støtter ikke HTTP-protokollversjonen som brukes i forespørselen.",
+  "506": "Transparent innholdsforhandling for forespørselen resulterer i en sirkulær referanse.",
+  "507": "Serveren er ikke i stand til å lagre representasjonen som trengs for å fullføre forespørselen.",
+  "508": "Serveren oppdaget en uendelig løkke mens den behandlet forespørselen.",
+  "510": "Ytterligere utvidelser til forespørselen er påkrevd for at serveren skal oppfylle den.",
+  "511": "Klienten må autentisere seg for å få nettverkstilgang."
+}
