+++ tools/http-status-code-lookup/descriptions/pl.json
@@ -1,0 +1,52 @@
+{
+  "100": "Serwer otrzymal naglowki zadan i klient powinien kontynuowac wysylanie ciala zadania.",
+  "101": "Zglaszajacy poprosil serwer o zmiane protokolow i serwer wyrazil zgode.",
+  "102": "Serwer otrzymal i przetwarza zadanie, ale odpowiedz nie jest jeszcze dostepna.",
+  "103": "Uzywane do zwracania niektórych naglowkow odpowiedzi przed ostateczna wiadomoscia HTTP.",
+  "200": "Zadanie powiodlo sie. Znaczenie sukcesu zalezy od metody HTTP.",
+  "201": "Zadanie zostalo spelnione i doprowadzilo do utworzenia nowego zasobu.",
+  "202": "Zadanie zostalo przyjete do przetworzenia, ale przetwarzanie nie zostalo zakonczone.",
+  "204": "Serwer pomyslnie przetworzyl zadanie i nie zwraca zadnej tresci.",
+  "206": "Serwer dostarcza tylko czesc zasobu z powodu naglowka zakresu wyslanego przez klienta.",
+  "300": "Zadanie ma wiecej niz jedna mozliwa odpowiedz. Agent uzytkownika powinien wybrac jedna z nich.",
+  "301": "URL zadanego zasobu zostal trwale zmieniony.",
+  "302": "URL zadanego zasobu zostal tymczasowo zmieniony.",
+  "304": "Wskazuje, ze zasob nie zostal zmodyfikowany od wersji okreslonej przez naglowki zadania.",
+  "307": "Serwer przekierowuje agenta uzytkownika do innego zasobu, jak wskazuje naglowek Location.",
+  "308": "Zasob jest teraz trwale zlokalizowany pod innym URL, okreslonym przez naglowek Location.",
+  "400": "Serwer nie moze lub nie bedzie przetwarzac zadania z powodu oczywistego bledu klienta.",
+  "401": "Uwierzytelnianie jest wymagane i nie powiodlo sie lub nie zostalo jeszcze dostarczone.",
+  "403": "Zadanie bylo wazne, ale serwer odmawia dzialania.",
+  "404": "Nie mozna znalezc zadanego zasobu, ale moze byc dostepny w przyszlosci.",
+  "405": "Metoda zadania nie jest obsługiwana dla zadanego zasobu.",
+  "406": "Zadany zasob jest w stanie generowac tylko tresc nieakceptowalna zgodnie z naglowkami Accept wyslanymi w zadaniu.",
+  "408": "Serwer osiagnal limit czasu podczas oczekiwania na zadanie.",
+  "409": "Nie mozna przetworzyc zadania z powodu konfliktu w zadaniu.",
+  "410": "Zadany zasob nie jest juz dostepny i nie bedzie juz dostepny.",
+  "411": "Zadanie nie okreslilo dlugosci swojej tresci, co jest wymagane przez zadany zasob.",
+  "412": "Serwer nie spelnia jednego z warunkow wstepnych, które zglaszajacy ustawil na zadaniu.",
+  "413": "Zadanie jest wieksze niz serwer jest gotowy lub zdolny przetworzyc.",
+  "414": "Podany URI byl zbyt dlugi, aby serwer mogl go przetworzyc.",
+  "415": "Jednostka zadania ma typ mediow, którego serwer lub zasob nie obsługuje.",
+  "416": "Klient poprosil o czesc pliku, ale serwer nie moze dostarczyc tej czesci.",
+  "417": "Serwer nie moze spelnic wymagan pola naglowka zadania Expect.",
+  "418": "Ten kod zostal zdefiniowany w 1998 roku jako jeden z tradycyjnych zartow primaaprilisowych IETF.",
+  "422": "Zadanie bylo dobrze sformulowane, ale nie mozna bylo mu sprostac z powodu bledow semantycznych.",
+  "425": "Wskazuje, ze serwer nie jest sklonny ryzykowac przetwarzania zadania, które moze zostac odtworzone.",
+  "426": "Klient powinien przejsc na inny protokol, taki jak TLS/1.0.",
+  "428": "Serwer zrodlowy wymaga, aby zadanie bylo warunkowe.",
+  "429": "Uzytkownik wyslal zbyt wiele zadan w danym okresie czasu.",
+  "431": "Serwer nie jest sklonny przetwarzac zadania, poniewaz jego pola naglowkow sa zbyt duze.",
+  "451": "Serwer odmawia dostepu do zasobu w konsekwencji zadanego prawnego.",
+  "500": "Ogolny komunikat bledu, podany, gdy napotkano nieoczekiwany warunek.",
+  "501": "Serwer nie rozpoznaje metody zadania lub brakuje mu zdolnosci do spelnienia zadania.",
+  "502": "Serwer dzialal jako brama lub proxy i otrzymal nieprawidlowa odpowiedz z serwera upstream.",
+  "503": "Serwer jest obecnie niedostepny (poniewaz jest przeciążony lub wylaczony do konserwacji).",
+  "504": "Serwer dzialal jako brama lub proxy i nie otrzymal odpowiedniej odpowiedzi z serwera upstream.",
+  "505": "Serwer nie obsługuje wersji protokolu HTTP uzywanej w zadaniu.",
+  "506": "Przezroczyste negocjacje tresci dla zadania skutkuja cyklicznym odniesieniem.",
+  "507": "Serwer nie jest w stanie przechowac reprezentacji potrzebnej do zakonczenia zadania.",
+  "508": "Serwer wykryl nieskonczona petle podczas przetwarzania zadania.",
+  "510": "Dalsze rozszerzenia zadania sa wymagane, aby serwer mogl je spelnic.",
+  "511": "Klient musi sie uwierzytelnic, aby uzyskac dostep do sieci."
+}
