+++ tools/http-status-code-lookup/descriptions/pt.json
@@ -1,0 +1,52 @@
+{
+  "100": "O servidor recebeu os cabecalhos da solicitacao e o cliente deve prosseguir para enviar o corpo da solicitacao.",
+  "101": "O solicitante pediu ao servidor para trocar protocolos e o servidor concordou.",
+  "102": "O servidor recebeu e esta processando a solicitacao, mas ainda nao ha resposta disponivel.",
+  "103": "Usado para retornar alguns cabecalhos de resposta antes da mensagem HTTP final.",
+  "200": "A solicitacao foi bem-sucedida. O significado do sucesso depende do metodo HTTP.",
+  "201": "A solicitacao foi atendida e resultou na criacao de um novo recurso.",
+  "202": "A solicitacao foi aceita para processamento, mas o processamento nao foi concluido.",
+  "204": "O servidor processou com sucesso a solicitacao e nao esta retornando nenhum conteudo.",
+  "206": "O servidor esta entregando apenas parte do recurso devido a um cabecalho de intervalo enviado pelo cliente.",
+  "300": "A solicitacao tem mais de uma resposta possivel. O agente do usuario deve escolher uma delas.",
+  "301": "A URL do recurso solicitado foi alterada permanentemente.",
+  "302": "A URL do recurso solicitado foi alterada temporariamente.",
+  "304": "Indica que o recurso nao foi modificado desde a versao especificada pelos cabecalhos da solicitacao.",
+  "307": "O servidor esta redirecionando o agente do usuario para um recurso diferente, conforme indicado por um cabecalho Location.",
+  "308": "O recurso agora esta permanentemente localizado em outra URL, especificada pelo cabecalho Location.",
+  "400": "O servidor nao pode ou nao processara a solicitacao devido a um erro aparente do cliente.",
+  "401": "A autenticacao e necessaria e falhou ou ainda nao foi fornecida.",
+  "403": "A solicitacao era valida, mas o servidor esta recusando a acao.",
+  "404": "O recurso solicitado nao foi encontrado, mas pode estar disponivel no futuro.",
+  "405": "Um metodo de solicitacao nao e suportado para o recurso solicitado.",
+  "406": "O recurso solicitado e capaz de gerar apenas conteudo nao aceitavel de acordo com os cabecalhos Accept enviados na solicitacao.",
+  "408": "O servidor atingiu o tempo limite aguardando a solicitacao.",
+  "409": "A solicitacao nao pôde ser processada devido a um conflito na solicitacao.",
+  "410": "O recurso solicitado nao esta mais disponivel e nao estara mais disponivel.",
+  "411": "A solicitacao nao especificou o comprimento de seu conteudo, que e exigido pelo recurso solicitado.",
+  "412": "O servidor nao atende a uma das pre-condicoes que o solicitante colocou na solicitacao.",
+  "413": "A solicitacao e maior do que o servidor esta disposto ou e capaz de processar.",
+  "414": "O URI fornecido era muito longo para o servidor processar.",
+  "415": "A entidade da solicitacao tem um tipo de midia que o servidor ou recurso nao suporta.",
+  "416": "O cliente pediu uma porcao do arquivo, mas o servidor nao pode fornecer essa porcao.",
+  "417": "O servidor nao pode atender aos requisitos do campo Expect do cabecalho da solicitacao.",
+  "418": "Este codigo foi definido em 1998 como uma das piadas tradicionais do Dia da Mentira da IETF.",
+  "422": "A solicitacao estava bem formada, mas nao pôde ser seguida devido a erros semanticos.",
+  "425": "Indica que o servidor nao esta disposto a arriscar processar uma solicitacao que pode ser reproduzida.",
+  "426": "O cliente deve mudar para um protocolo diferente, como TLS/1.0.",
+  "428": "O servidor de origem requer que a solicitacao seja condicional.",
+  "429": "O usuario enviou muitas solicitacoes em um determinado periodo de tempo.",
+  "431": "O servidor nao esta disposto a processar a solicitacao porque seus campos de cabecalho sao muito grandes.",
+  "451": "O servidor esta negando acesso ao recurso como consequencia de uma demanda legal.",
+  "500": "Uma mensagem de erro generica, dada quando uma condicao inesperada foi encontrada.",
+  "501": "O servidor nao reconhece o metodo de solicitacao ou carece da capacidade de atender a solicitacao.",
+  "502": "O servidor estava atuando como um gateway ou proxy e recebeu uma resposta invalida do servidor upstream.",
+  "503": "O servidor nao esta disponivel no momento (porque esta sobrecarregado ou em manutencao).",
+  "504": "O servidor estava atuando como um gateway ou proxy e nao recebeu uma resposta oportuna do servidor upstream.",
+  "505": "O servidor nao suporta a versao do protocolo HTTP usada na solicitacao.",
+  "506": "A negociacao transparente de conteudo para a solicitacao resulta em uma referencia circular.",
+  "507": "O servidor e incapaz de armazenar a representacao necessaria para completar a solicitacao.",
+  "508": "O servidor detectou um loop infinito ao processar a solicitacao.",
+  "510": "Sao necessarias extensoes adicionais a solicitacao para que o servidor a atenda.",
+  "511": "O cliente precisa se autenticar para obter acesso a rede."
+}
