+++ tools/http-status-code-lookup/descriptions/ru.json
@@ -1,0 +1,52 @@
+{
+  "100": "Сервер получил заголовки запроса, и клиент должен продолжить отправку тела запроса.",
+  "101": "Запрашивающая сторона попросила сервер переключить протоколы, и сервер согласился.",
+  "102": "Сервер получил и обрабатывает запрос, но ответ еще недоступен.",
+  "103": "Используется для возврата некоторых заголовков ответа перед окончательным HTTP-сообщением.",
+  "200": "Запрос успешен. Значение успеха зависит от HTTP-метода.",
+  "201": "Запрос выполнен и привел к созданию нового ресурса.",
+  "202": "Запрос принят для обработки, но обработка еще не завершена.",
+  "204": "Сервер успешно обработал запрос и не возвращает никакого содержимого.",
+  "206": "Сервер доставляет только часть ресурса из-за заголовка диапазона, отправленного клиентом.",
+  "300": "Запрос имеет более одного возможного ответа. Агент пользователя должен выбрать один из них.",
+  "301": "URL запрошенного ресурса был изменен навсегда.",
+  "302": "URL запрошенного ресурса был изменен временно.",
+  "304": "Указывает, что ресурс не был изменен с версии, указанной в заголовках запроса.",
+  "307": "Сервер перенаправляет агента пользователя на другой ресурс, как указано в заголовке Location.",
+  "308": "Ресурс теперь постоянно находится по другому URL, указанному в заголовке Location.",
+  "400": "Сервер не может или не будет обрабатывать запрос из-за очевидной ошибки клиента.",
+  "401": "Требуется аутентификация, и она не удалась или еще не предоставлена.",
+  "403": "Запрос был действительным, но сервер отказывается выполнять действие.",
+  "404": "Запрошенный ресурс не найден, но может быть доступен в будущем.",
+  "405": "Метод запроса не поддерживается для запрошенного ресурса.",
+  "406": "Запрошенный ресурс способен генерировать только содержимое, неприемлемое согласно заголовкам Accept, отправленным в запросе.",
+  "408": "Сервер превысил время ожидания запроса.",
+  "409": "Запрос не мог быть обработан из-за конфликта в запросе.",
+  "410": "Запрошенный ресурс больше недоступен и больше не будет доступен.",
+  "411": "Запрос не указал длину своего содержимого, что требуется запрошенным ресурсом.",
+  "412": "Сервер не соответствует одному из предварительных условий, которые запрашивающая сторона установила для запроса.",
+  "413": "Запрос больше, чем сервер готов или способен обработать.",
+  "414": "Предоставленный URI был слишком длинным для обработки сервером.",
+  "415": "Сущность запроса имеет тип медиа, который сервер или ресурс не поддерживает.",
+  "416": "Клиент запросил часть файла, но сервер не может предоставить эту часть.",
+  "417": "Сервер не может соответствовать требованиям поля Expect заголовка запроса.",
+  "418": "Этот код был определен в 1998 году как одна из традиционных шуток IETF на 1 апреля.",
+  "422": "Запрос был правильно сформирован, но не мог быть выполнен из-за семантических ошибок.",
+  "425": "Указывает, что сервер не желает рисковать обработкой запроса, который может быть воспроизведен.",
+  "426": "Клиент должен переключиться на другой протокол, такой как TLS/1.0.",
+  "428": "Исходный сервер требует, чтобы запрос был условным.",
+  "429": "Пользователь отправил слишком много запросов за определенный период времени.",
+  "431": "Сервер не желает обрабатывать запрос, потому что его поля заголовков слишком большие.",
+  "451": "Сервер отказывает в доступе к ресурсу в результате правового требования.",
+  "500": "Общее сообщение об ошибке, выдаваемое при возникновении неожиданного условия.",
+  "501": "Сервер либо не распознает метод запроса, либо не имеет возможности выполнить запрос.",
+  "502": "Сервер действовал как шлюз или прокси и получил неверный ответ от вышестоящего сервера.",
+  "503": "Сервер в настоящее время недоступен (потому что он перегружен или находится на обслуживании).",
+  "504": "Сервер действовал как шлюз или прокси и не получил своевременный ответ от вышестоящего сервера.",
+  "505": "Сервер не поддерживает версию протокола HTTP, используемую в запросе.",
+  "506": "Прозрачное согласование содержимого для запроса приводит к циклической ссылке.",
+  "507": "Сервер не может сохранить представление, необходимое для завершения запроса.",
+  "508": "Сервер обнаружил бесконечный цикл при обработке запроса.",
+  "510": "Для выполнения запроса сервером требуются дальнейшие расширения запроса.",
+  "511": "Клиенту необходимо пройти аутентификацию для получения доступа к сети."
+}
