+++ tools/http-status-code-lookup/descriptions/sv.json
@@ -1,0 +1,52 @@
+{
+  "100": "Servern har mottagit begäranhuvudena och klienten bör fortsätta att skicka begärandekroppen.",
+  "101": "Begäraren har bett servern att byta protokoll och servern har gått med på det.",
+  "102": "Servern har mottagit och bearbetar begäran, men inget svar är tillgängligt ännu.",
+  "103": "Används för att returnera vissa svarshuvuden före det slutliga HTTP-meddelandet.",
+  "200": "Begäran lyckades. Betydelsen av framgång beror på HTTP-metoden.",
+  "201": "Begäran har uppfyllts och resulterade i att en ny resurs skapades.",
+  "202": "Begäran har accepterats för bearbetning, men bearbetningen har inte slutförts.",
+  "204": "Servern bearbetade begäran framgångsrikt och returnerar inget innehåll.",
+  "206": "Servern levererar endast en del av resursen på grund av ett intervallhuvud som skickats av klienten.",
+  "300": "Begäran har mer än ett möjligt svar. Användaragenten bör välja ett av dem.",
+  "301": "URL:en för den begärda resursen har ändrats permanent.",
+  "302": "URL:en för den begärda resursen har ändrats tillfälligt.",
+  "304": "Anger att resursen inte har ändrats sedan versionen som anges av begäranhuvudena.",
+  "307": "Servern omdirigerar användaragenten till en annan resurs, som anges av ett Location-huvud.",
+  "308": "Resursen finns nu permanent på en annan URL, angiven av Location-huvudet.",
+  "400": "Servern kan inte eller kommer inte att bearbeta begäran på grund av ett uppenbart klientfel.",
+  "401": "Autentisering krävs och har misslyckats eller har inte ännu tillhandahållits.",
+  "403": "Begäran var giltig, men servern vägrar att agera.",
+  "404": "Den begärda resursen kunde inte hittas men kan vara tillgänglig i framtiden.",
+  "405": "En begäranmetod stöds inte för den begärda resursen.",
+  "406": "Den begärda resursen kan endast generera innehåll som inte är acceptabelt enligt Accept-huvudena som skickades i begäran.",
+  "408": "Servern nådde timeout medan den väntade på begäran.",
+  "409": "Begäran kunde inte bearbetas på grund av en konflikt i begäran.",
+  "410": "Den begärda resursen är inte längre tillgänglig och kommer inte att vara tillgänglig igen.",
+  "411": "Begäran specificerade inte längden på sitt innehåll, vilket krävs av den begärda resursen.",
+  "412": "Servern uppfyller inte en av förutsättningarna som begäraren ställde på begäran.",
+  "413": "Begäran är större än servern är villig eller kan bearbeta.",
+  "414": "Den angivna URI:n var för lång för att servern skulle kunna bearbeta den.",
+  "415": "Begäranenheten har en mediatyp som servern eller resursen inte stöder.",
+  "416": "Klienten har bett om en del av filen, men servern kan inte leverera den delen.",
+  "417": "Servern kan inte uppfylla kraven för Expect-begäranhuvudfältet.",
+  "418": "Denna kod definierades 1998 som en av de traditionella IETF aprilskämten.",
+  "422": "Begäran var välformad men kunde inte följas på grund av semantiska fel.",
+  "425": "Anger att servern inte är villig att riskera att bearbeta en begäran som kan spelas upp igen.",
+  "426": "Klienten bör växla till ett annat protokoll som TLS/1.0.",
+  "428": "Ursprungsservern kräver att begäran är villkorad.",
+  "429": "Användaren har skickat för många begäranden under en given tidsperiod.",
+  "431": "Servern är inte villig att bearbeta begäran eftersom dess huvudfält är för stora.",
+  "451": "Servern nekar åtkomst till resursen som en följd av en juridisk begäran.",
+  "500": "Ett generiskt felmeddelande, ges när ett oväntat tillstånd påträffades.",
+  "501": "Servern känner inte igen begäranmetoden eller saknar förmågan att uppfylla begäran.",
+  "502": "Servern fungerade som en gateway eller proxy och fick ett ogiltigt svar från upstream-servern.",
+  "503": "Servern är för närvarande inte tillgänglig (eftersom den är överbelastad eller nere för underhåll).",
+  "504": "Servern fungerade som en gateway eller proxy och fick inte ett svar i tid från upstream-servern.",
+  "505": "Servern stöder inte HTTP-protokollversionen som används i begäran.",
+  "506": "Transparent innehållsförhandling för begäran resulterar i en cirkulär referens.",
+  "507": "Servern kan inte lagra representationen som behövs för att slutföra begäran.",
+  "508": "Servern upptäckte en oändlig loop medan den bearbetade begäran.",
+  "510": "Ytterligare utökningar av begäran krävs för att servern ska kunna uppfylla den.",
+  "511": "Klienten behöver autentisera sig för att få nätverksåtkomst."
+}
