+++ tools/http-status-code-lookup/descriptions/th.json
@@ -1,0 +1,52 @@
+{
+  "100": "เซิร์ฟเวอร์ได้รับส่วนหัวของคำขอแล้ว และไคลเอนต์ควรดำเนินการส่งเนื้อหาของคำขอต่อไป",
+  "101": "ผู้ร้องขอได้ขอให้เซิร์ฟเวอร์เปลี่ยนโปรโตคอล และเซิร์ฟเวอร์ได้ตกลง",
+  "102": "เซิร์ฟเวอร์ได้รับและกำลังประมวลผลคำขอ แต่ยังไม่มีคำตอบ",
+  "103": "ใช้เพื่อส่งคืนส่วนหัวของคำตอบบางส่วนก่อนข้อความ HTTP สุดท้าย",
+  "200": "คำขอนั้นสำเร็จ ความหมายของความสำเร็จขึ้นอยู่กับวิธี HTTP",
+  "201": "คำขอได้รับการปฏิบัติและส่งผลให้มีการสร้างทรัพยากรใหม่",
+  "202": "คำขอได้รับการยอมรับสำหรับการประมวลผล แต่การประมวลผลยังไม่เสร็จสมบูรณ์",
+  "204": "เซิร์ฟเวอร์ประมวลผลคำขอสำเร็จและไม่ส่งคืนเนื้อหาใดๆ",
+  "206": "เซิร์ฟเวอร์กำลังส่งมอบทรัพยากรเพียงบางส่วนเนื่องจากส่วนหัวของช่วงที่ส่งโดยไคลเอนต์",
+  "300": "คำขอมีคำตอบที่เป็นไปได้มากกว่าหนึ่งคำตอบ ตัวแทนผู้ใช้ควรเลือกหนึ่งในนั้น",
+  "301": "URL ของทรัพยากรที่ร้องขอได้รับการเปลี่ยนแปลงอย่างถาวร",
+  "302": "URL ของทรัพยากรที่ร้องขอได้รับการเปลี่ยนแปลงชั่วคราว",
+  "304": "ระบุว่าทรัพยากรไม่ได้รับการแก้ไขตั้งแต่เวอร์ชันที่ระบุโดยส่วนหัวของคำขอ",
+  "307": "เซิร์ฟเวอร์กำลังเปลี่ยนเส้นทางตัวแทนผู้ใช้ไปยังทรัพยากรอื่น ตามที่ระบุโดยส่วนหัว Location",
+  "308": "ทรัพยากรอยู่ที่ URL อื่นอย่างถาวร ตามที่ระบุโดยส่วนหัว Location",
+  "400": "เซิร์ฟเวอร์ไม่สามารถหรือจะไม่ประมวลผลคำขอเนื่องจากข้อผิดพลาดของไคลเอนต์ที่ชัดเจน",
+  "401": "การรับรองความถูกต้องจำเป็นและล้มเหลวหรือยังไม่ได้รับ",
+  "403": "คำขอถูกต้อง แต่เซิร์ฟเวอร์ปฏิเสธการดำเนินการ",
+  "404": "ไม่พบทรัพยากรที่ร้องขอ แต่อาจมีให้ในอนาคต",
+  "405": "วิธีการร้องขอไม่ได้รับการสนับสนุนสำหรับทรัพยากรที่ร้องขอ",
+  "406": "ทรัพยากรที่ร้องขอสามารถสร้างเนื้อหาได้เฉพาะที่ไม่ยอมรับตามส่วนหัว Accept ที่ส่งในคำขอ",
+  "408": "เซิร์ฟเวอร์หมดเวลารอคำขอ",
+  "409": "ไม่สามารถประมวลผลคำขอได้เนื่องจากความขัดแย้งในคำขอ",
+  "410": "ทรัพยากรที่ร้องขอไม่สามารถใช้ได้อีกต่อไปและจะไม่สามารถใช้ได้อีก",
+  "411": "คำขอไม่ได้ระบุความยาวของเนื้อหา ซึ่งจำเป็นสำหรับทรัพยากรที่ร้องขอ",
+  "412": "เซิร์ฟเวอร์ไม่เป็นไปตามเงื่อนไขเบื้องต้นข้อใดข้อหนึ่งที่ผู้ร้องขอกำหนดไว้ในคำขอ",
+  "413": "คำขอมีขนาดใหญ่กว่าเซิร์ฟเวอร์ที่ยินดีหรือสามารถประมวลผลได้",
+  "414": "URI ที่ให้มามีความยาวเกินไปสำหรับเซิร์ฟเวอร์ที่จะประมวลผล",
+  "415": "เอนทิตีของคำขอมีประเภทสื่อที่เซิร์ฟเวอร์หรือทรัพยากรไม่รองรับ",
+  "416": "ไคลเอนต์ได้ขอส่วนหนึ่งของไฟล์ แต่เซิร์ฟเวอร์ไม่สามารถจัดหาส่วนนั้นได้",
+  "417": "เซิร์ฟเวอร์ไม่สามารถเป็นไปตามข้อกำหนดของฟิลด์ส่วนหัวคำขอ Expect",
+  "418": "รหัสนี้ถูกกำหนดในปี 1998 เป็นหนึ่งในมุกตลกวันเมษาหน้าโง่แบบดั้งเดิมของ IETF",
+  "422": "คำขอถูกจัดรูปแบบอย่างดี แต่ไม่สามารถปฏิบัติตามได้เนื่องจากข้อผิดพลาดทางความหมาย",
+  "425": "ระบุว่าเซิร์ฟเวอร์ไม่เต็มใจที่จะเสี่ยงกับการประมวลผลคำขอที่อาจถูกเล่นซ้ำ",
+  "426": "ไคลเอนต์ควรเปลี่ยนไปใช้โปรโตคอลอื่น เช่น TLS/1.0",
+  "428": "เซิร์ฟเวอร์ต้นทางต้องการให้คำขอมีเงื่อนไข",
+  "429": "ผู้ใช้ได้ส่งคำขอมากเกินไปในช่วงเวลาที่กำหนด",
+  "431": "เซิร์ฟเวอร์ไม่เต็มใจที่จะประมวลผลคำขอเนื่องจากฟิลด์ส่วนหัวมีขนาดใหญ่เกินไป",
+  "451": "เซิร์ฟเวอร์ปฏิเสธการเข้าถึงทรัพยากรอันเป็นผลมาจากคำขอทางกฎหมาย",
+  "500": "ข้อความแสดงข้อผิดพลาดทั่วไป กำหนดเมื่อพบเงื่อนไขที่ไม่คาดคิด",
+  "501": "เซิร์ฟเวอร์ไม่รู้จักวิธีการร้องขอ หรือขาดความสามารถในการปฏิบัติตามคำขอ",
+  "502": "เซิร์ฟเวอร์ทำหน้าที่เป็นเกตเวย์หรือพร็อกซี่ และได้รับคำตอบที่ไม่ถูกต้องจากเซิร์ฟเวอร์ต้นทาง",
+  "503": "เซิร์ฟเวอร์ไม่พร้อมใช้งานในขณะนี้ (เนื่องจากโอเวอร์โหลดหรือปิดเพื่อการบำรุงรักษา)",
+  "504": "เซิร์ฟเวอร์ทำหน้าที่เป็นเกตเวย์หรือพร็อกซี่ และไม่ได้รับคำตอบทันเวลาจากเซิร์ฟเวอร์ต้นทาง",
+  "505": "เซิร์ฟเวอร์ไม่รองรับเวอร์ชันโปรโตคอล HTTP ที่ใช้ในคำขอ",
+  "506": "การเจรจาเนื้อหาอย่างโปร่งใสสำหรับคำขอส่งผลให้เกิดการอ้างอิงแบบวงกลม",
+  "507": "เซิร์ฟเวอร์ไม่สามารถจัดเก็บการแสดงผลที่จำเป็นเพื่อทำให้คำขอสมบูรณ์",
+  "508": "เซิร์ฟเวอร์ตรวจพบลูปไม่สิ้นสุดขณะประมวลผลคำขอ",
+  "510": "ต้องมีการขยายคำขอเพิ่มเติมเพื่อให้เซิร์ฟเวอร์สามารถปฏิบัติตามได้",
+  "511": "ไคลเอนต์ต้องรับรองความถูกต้องเพื่อให้ได้การเข้าถึงเครือข่าย"
+}
