+++ tools/http-status-code-lookup/descriptions/tr.json
@@ -1,0 +1,52 @@
+{
+  "100": "Sunucu istek basliklarini aldi ve istemci istek gövdesini göndermeye devam etmelidir.",
+  "101": "Istek sahibi sunucudan protokolleri degistirmesini istedi ve sunucu kabul etti.",
+  "102": "Sunucu istegi aldi ve isliyor, ancak henüz yanit mevcut degil.",
+  "103": "Son HTTP mesajindan önce bazi yanit basliklarini döndürmek için kullanilir.",
+  "200": "Istek basarili oldu. Basarinin anlami HTTP yöntemine baglidir.",
+  "201": "Istek tamamlandi ve yeni bir kaynak olusturulmasina neden oldu.",
+  "202": "Istek isleme için kabul edildi, ancak isleme tamamlanmadi.",
+  "204": "Sunucu istegi basariyla isledi ve herhangi bir içerik döndürmüyor.",
+  "206": "Sunucu, istemci tarafindan gönderilen bir aralik basligi nedeniyle kaynagin yalnizca bir kismini teslim ediyor.",
+  "300": "Istegin birden fazla olası yaniti var. Kullanici aracisi bunlardan birini seçmelidir.",
+  "301": "Istek edilen kaynagin URL'si kalici olarak degistirildi.",
+  "302": "Istek edilen kaynagin URL'si geçici olarak degistirildi.",
+  "304": "Kaynagin, istek basliklari tarafindan belirtilen sürümden bu yana degistirilmedigini gösterir.",
+  "307": "Sunucu, bir Location basligiyla belirtildigi gibi kullanici aracisini farkli bir kaynaga yönlendiriyor.",
+  "308": "Kaynak artik kalici olarak Location basligiyla belirtilen baska bir URL'de bulunuyor.",
+  "400": "Sunucu, belirgin bir istemci hatasi nedeniyle istegi isleyemez veya islemeyecektir.",
+  "401": "Kimlik dogrulama gereklidir ve basarisiz oldu veya henüz saglanmadi.",
+  "403": "Istek geçerliydi, ancak sunucu eylemi reddediyor.",
+  "404": "Istek edilen kaynak bulunamadi ancak gelecekte kullanilabilir olabilir.",
+  "405": "Istek edilen kaynak için bir istek yöntemi desteklenmiyor.",
+  "406": "Istek edilen kaynak, istekte gönderilen Accept basliklarina göre kabul edilemez içerik üretebilir.",
+  "408": "Sunucu istegi beklerken zaman asimina ugradi.",
+  "409": "Istekteki bir çakisma nedeniyle istek islenemedi.",
+  "410": "Istek edilen kaynak artik kullanilabilir degil ve tekrar kullanilabilir olmayacak.",
+  "411": "Istek, içeriginin uzunlugunu belirtmedi, bu da istenen kaynak tarafindan gereklidir.",
+  "412": "Sunucu, istek sahibinin istege koydugu ön kosullardan birini karsilamiyor.",
+  "413": "Istek, sunucunun islemeye hazir oldugu veya isleyebilecegi boyuttan daha büyüktür.",
+  "414": "Saglanan URI sunucunun islemesi için çok uzundu.",
+  "415": "Istek varliginin sunucu veya kaynak tarafindan desteklenmeyen bir medya türü var.",
+  "416": "Istemci dosyanin bir kismini istedi, ancak sunucu bu kismi saglayamaz.",
+  "417": "Sunucu, istek baslik alani Expect'in gereksinimlerini karsilayamaz.",
+  "418": "Bu kod 1998'de geleneksel IETF Nisan 1 saka saka'larindan biri olarak tanimlandi.",
+  "422": "Istek iyi biçimlendirilmisti ancak anlamsal hatalar nedeniyle takip edilemedi.",
+  "425": "Sunucunun yeniden oynatilabilecek bir istegi isleme riskini almak istemedigini gösterir.",
+  "426": "Istemci TLS/1.0 gibi farkli bir protokole geçmelidir.",
+  "428": "Kaynak sunucu istegin kosullu olmasini gerektirir.",
+  "429": "Kullanici belirli bir süre içinde çok fazla istek gönderdi.",
+  "431": "Sunucu, baslik alanlari çok büyük oldugu için istegi islemek istemiyor.",
+  "451": "Sunucu, yasal bir talep sonucunda kaynaga erisimi reddediyor.",
+  "500": "Beklenmeyen bir durumla karsilasildiginda verilen genel bir hata mesaji.",
+  "501": "Sunucu istek yöntemini tanimiyor veya istegi yerine getirme yeteneginden yoksun.",
+  "502": "Sunucu bir ağ geçidi veya proxy olarak çalisiyordu ve yukari akis sunucusundan geçersiz bir yanit aldi.",
+  "503": "Sunucu su anda kullanilabilir degil (asiri yüklü oldugu veya bakim için kapali oldugu için).",
+  "504": "Sunucu bir ağ geçidi veya proxy olarak çalisiyordu ve yukari akis sunucusundan zamaninda bir yanit aldi.",
+  "505": "Sunucu istekte kullanilan HTTP protokol sürümünü desteklemiyor.",
+  "506": "Istek için saydam içerik müzakere döngüsel bir referansla sonuçlanir.",
+  "507": "Sunucu istegi tamamlamak için gereken temsili depolayamaz.",
+  "508": "Sunucu istegi islerken sonsuz bir döngü tespit etti.",
+  "510": "Sunucunun bunu yerine getirmesi için istege daha fazla uzanti gereklidir.",
+  "511": "Istemcinin ağ erisimi kazanmak için kimlik dogrulamasi yapmasi gerekiyor."
+}
