+++ tools/http-status-code-lookup/descriptions/vi.json
@@ -1,0 +1,52 @@
+{
+  "100": "Máy chủ đã nhận được các tiêu đề yêu cầu và máy khách nên tiếp tục gửi phần thân yêu cầu.",
+  "101": "Người yêu cầu đã yêu cầu máy chủ chuyển đổi giao thức và máy chủ đã đồng ý.",
+  "102": "Máy chủ đã nhận và đang xử lý yêu cầu, nhưng chưa có phản hồi.",
+  "103": "Được sử dụng để trả về một số tiêu đề phản hồi trước thông điệp HTTP cuối cùng.",
+  "200": "Yêu cầu đã thành công. Ý nghĩa của thành công phụ thuộc vào phương thức HTTP.",
+  "201": "Yêu cầu đã được thực hiện và dẫn đến việc tạo một tài nguyên mới.",
+  "202": "Yêu cầu đã được chấp nhận để xử lý, nhưng việc xử lý chưa hoàn thành.",
+  "204": "Máy chủ đã xử lý thành công yêu cầu và không trả về nội dung nào.",
+  "206": "Máy chủ chỉ cung cấp một phần tài nguyên do tiêu đề phạm vi được gửi bởi máy khách.",
+  "300": "Yêu cầu có nhiều hơn một phản hồi có thể. Tác nhân người dùng nên chọn một trong số chúng.",
+  "301": "URL của tài nguyên được yêu cầu đã được thay đổi vĩnh viễn.",
+  "302": "URL của tài nguyên được yêu cầu đã được thay đổi tạm thời.",
+  "304": "Chỉ ra rằng tài nguyên chưa được sửa đổi kể từ phiên bản được chỉ định bởi các tiêu đề yêu cầu.",
+  "307": "Máy chủ đang chuyển hướng tác nhân người dùng đến một tài nguyên khác, như được chỉ định bởi tiêu đề Location.",
+  "308": "Tài nguyên hiện đang được đặt vĩnh viễn tại một URL khác, được chỉ định bởi tiêu đề Location.",
+  "400": "Máy chủ không thể hoặc sẽ không xử lý yêu cầu do lỗi rõ ràng của máy khách.",
+  "401": "Xác thực là bắt buộc và đã thất bại hoặc chưa được cung cấp.",
+  "403": "Yêu cầu hợp lệ, nhưng máy chủ từ chối thực hiện hành động.",
+  "404": "Không tìm thấy tài nguyên được yêu cầu nhưng có thể có sẵn trong tương lai.",
+  "405": "Phương thức yêu cầu không được hỗ trợ cho tài nguyên được yêu cầu.",
+  "406": "Tài nguyên được yêu cầu chỉ có thể tạo nội dung không chấp nhận được theo các tiêu đề Accept được gửi trong yêu cầu.",
+  "408": "Máy chủ đã hết thời gian chờ yêu cầu.",
+  "409": "Không thể xử lý yêu cầu do xung đột trong yêu cầu.",
+  "410": "Tài nguyên được yêu cầu không còn khả dụng và sẽ không khả dụng nữa.",
+  "411": "Yêu cầu không chỉ định độ dài nội dung của nó, điều này được yêu cầu bởi tài nguyên được yêu cầu.",
+  "412": "Máy chủ không đáp ứng một trong các điều kiện tiên quyết mà người yêu cầu đặt trên yêu cầu.",
+  "413": "Yêu cầu lớn hơn mức máy chủ sẵn sàng hoặc có thể xử lý.",
+  "414": "URI được cung cấp quá dài để máy chủ xử lý.",
+  "415": "Thực thể yêu cầu có loại phương tiện mà máy chủ hoặc tài nguyên không hỗ trợ.",
+  "416": "Máy khách đã yêu cầu một phần của tệp, nhưng máy chủ không thể cung cấp phần đó.",
+  "417": "Máy chủ không thể đáp ứng các yêu cầu của trường tiêu đề yêu cầu Expect.",
+  "418": "Mã này được định nghĩa vào năm 1998 như một trong những trò đùa truyền thống của IETF vào ngày Cá tháng Tư.",
+  "422": "Yêu cầu được định dạng tốt nhưng không thể được thực hiện do lỗi ngữ nghĩa.",
+  "425": "Chỉ ra rằng máy chủ không sẵn sàng mạo hiểm xử lý một yêu cầu có thể được phát lại.",
+  "426": "Máy khách nên chuyển sang một giao thức khác như TLS/1.0.",
+  "428": "Máy chủ gốc yêu cầu yêu cầu phải có điều kiện.",
+  "429": "Người dùng đã gửi quá nhiều yêu cầu trong một khoảng thời gian nhất định.",
+  "431": "Máy chủ không sẵn sàng xử lý yêu cầu vì các trường tiêu đề của nó quá lớn.",
+  "451": "Máy chủ đang từ chối quyền truy cập vào tài nguyên như một hậu quả của yêu cầu pháp lý.",
+  "500": "Thông báo lỗi chung, được đưa ra khi gặp phải điều kiện không mong đợi.",
+  "501": "Máy chủ không nhận ra phương thức yêu cầu hoặc thiếu khả năng thực hiện yêu cầu.",
+  "502": "Máy chủ đang hoạt động như một cổng hoặc proxy và nhận được phản hồi không hợp lệ từ máy chủ upstream.",
+  "503": "Máy chủ hiện không khả dụng (vì quá tải hoặc đang bảo trì).",
+  "504": "Máy chủ đang hoạt động như một cổng hoặc proxy và không nhận được phản hồi kịp thời từ máy chủ upstream.",
+  "505": "Máy chủ không hỗ trợ phiên bản giao thức HTTP được sử dụng trong yêu cầu.",
+  "506": "Thương lượng nội dung minh bạch cho yêu cầu dẫn đến tham chiếu vòng tròn.",
+  "507": "Máy chủ không thể lưu trữ biểu diễn cần thiết để hoàn thành yêu cầu.",
+  "508": "Máy chủ đã phát hiện một vòng lặp vô hạn trong khi xử lý yêu cầu.",
+  "510": "Cần mở rộng thêm yêu cầu để máy chủ thực hiện nó.",
+  "511": "Máy khách cần xác thực để có quyền truy cập mạng."
+}
