+++ tools/http-status-code-lookup/descriptions/zh-CN.json
@@ -1,0 +1,52 @@
+{
+  "100": "服务器已收到请求头，客户端应继续发送请求体。",
+  "101": "请求者已要求服务器切换协议，服务器已同意。",
+  "102": "服务器已收到并正在处理请求，但尚未有响应可用。",
+  "103": "用于在最终 HTTP 消息之前返回一些响应头。",
+  "200": "请求已成功。成功的含义取决于 HTTP 方法。",
+  "201": "请求已完成，并导致创建了新资源。",
+  "202": "请求已被接受处理，但处理尚未完成。",
+  "204": "服务器成功处理了请求，但不返回任何内容。",
+  "206": "由于客户端发送的范围头，服务器仅提供资源的一部分。",
+  "300": "请求有多个可能的响应。用户代理应选择其中一个。",
+  "301": "请求资源的 URL 已永久更改。",
+  "302": "请求资源的 URL 已临时更改。",
+  "304": "表示自请求头指定的版本以来，资源未被修改。",
+  "307": "服务器正在将用户代理重定向到不同的资源，如 Location 头所示。",
+  "308": "资源现在永久位于另一个 URL，由 Location 头指定。",
+  "400": "由于明显的客户端错误，服务器无法或不会处理请求。",
+  "401": "需要身份验证，但已失败或尚未提供。",
+  "403": "请求有效，但服务器拒绝执行操作。",
+  "404": "找不到请求的资源，但将来可能可用。",
+  "405": "请求的资源不支持该请求方法。",
+  "406": "请求的资源只能生成根据请求中发送的 Accept 头不可接受的内容。",
+  "408": "服务器等待请求超时。",
+  "409": "由于请求中的冲突，无法处理请求。",
+  "410": "请求的资源不再可用，并且不会再可用。",
+  "411": "请求未指定其内容的长度，这是请求资源所需的。",
+  "412": "服务器不满足请求者对请求设置的前提条件之一。",
+  "413": "请求大于服务器愿意或能够处理的大小。",
+  "414": "提供的 URI 太长，服务器无法处理。",
+  "415": "请求实体具有服务器或资源不支持的媒体类型。",
+  "416": "客户端请求文件的一部分，但服务器无法提供该部分。",
+  "417": "服务器无法满足 Expect 请求头字段的要求。",
+  "418": "此代码在 1998 年定义为传统的 IETF 愚人节笑话之一。",
+  "422": "请求格式正确，但由于语义错误而无法执行。",
+  "425": "表示服务器不愿意冒险处理可能被重放的请求。",
+  "426": "客户端应切换到不同的协议，例如 TLS/1.0。",
+  "428": "源服务器要求请求是有条件的。",
+  "429": "用户在给定时间内发送了太多请求。",
+  "431": "服务器不愿意处理请求，因为其头字段太大。",
+  "451": "由于法律要求，服务器拒绝访问资源。",
+  "500": "通用错误消息，在遇到意外情况时给出。",
+  "501": "服务器无法识别请求方法，或缺乏满足请求的能力。",
+  "502": "服务器充当网关或代理，并从上游服务器收到无效响应。",
+  "503": "服务器当前不可用（因为过载或维护中）。",
+  "504": "服务器充当网关或代理，但未从上游服务器收到及时响应。",
+  "505": "服务器不支持请求中使用的 HTTP 协议版本。",
+  "506": "请求的透明内容协商导致循环引用。",
+  "507": "服务器无法存储完成请求所需的表示。",
+  "508": "服务器在处理请求时检测到无限循环。",
+  "510": "需要进一步扩展请求，服务器才能完成它。",
+  "511": "客户端需要身份验证才能获得网络访问权限。"
+}
