+++ tools/http-status-code-lookup/descriptions/zh-TW.json
@@ -1,0 +1,52 @@
+{
+  "100": "伺服器已收到請求標頭，用戶端應繼續發送請求體。",
+  "101": "請求者已要求伺服器切換協議，伺服器已同意。",
+  "102": "伺服器已收到並正在處理請求，但尚未有回應可用。",
+  "103": "用於在最終 HTTP 訊息之前返回一些回應標頭。",
+  "200": "請求已成功。成功的含義取決於 HTTP 方法。",
+  "201": "請求已完成，並導致創建了新資源。",
+  "202": "請求已被接受處理，但處理尚未完成。",
+  "204": "伺服器成功處理了請求，但不返回任何內容。",
+  "206": "由於用戶端發送的範圍標頭，伺服器僅提供資源的一部分。",
+  "300": "請求有多個可能的回應。用戶代理應選擇其中一個。",
+  "301": "請求資源的 URL 已永久更改。",
+  "302": "請求資源的 URL 已臨時更改。",
+  "304": "表示自請求標頭指定的版本以來，資源未被修改。",
+  "307": "伺服器正在將用戶代理重新導向到不同的資源，如 Location 標頭所示。",
+  "308": "資源現在永久位於另一個 URL，由 Location 標頭指定。",
+  "400": "由於明顯的用戶端錯誤，伺服器無法或不會處理請求。",
+  "401": "需要身份驗證，但已失敗或尚未提供。",
+  "403": "請求有效，但伺服器拒絕執行操作。",
+  "404": "找不到請求的資源，但將來可能可用。",
+  "405": "請求的資源不支持該請求方法。",
+  "406": "請求的資源只能生成根據請求中發送的 Accept 標頭不可接受的內容。",
+  "408": "伺服器等待請求超時。",
+  "409": "由於請求中的衝突，無法處理請求。",
+  "410": "請求的資源不再可用，並且不會再可用。",
+  "411": "請求未指定其內容的長度，這是請求資源所需的。",
+  "412": "伺服器不滿足請求者對請求設置的前提條件之一。",
+  "413": "請求大於伺服器願意或能夠處理的大小。",
+  "414": "提供的 URI 太長，伺服器無法處理。",
+  "415": "請求實體具有伺服器或資源不支持的媒體類型。",
+  "416": "用戶端請求檔案的一部分，但伺服器無法提供該部分。",
+  "417": "伺服器無法滿足 Expect 請求標頭欄位的要求。",
+  "418": "此代碼在 1998 年定義為傳統的 IETF 愚人節笑話之一。",
+  "422": "請求格式正確，但由於語義錯誤而無法執行。",
+  "425": "表示伺服器不願意冒險處理可能被重放的請求。",
+  "426": "用戶端應切換到不同的協議，例如 TLS/1.0。",
+  "428": "源伺服器要求請求是有條件的。",
+  "429": "用戶在給定時間內發送了太多請求。",
+  "431": "伺服器不願意處理請求，因為其標頭欄位太大。",
+  "451": "由於法律要求，伺服器拒絕訪問資源。",
+  "500": "通用錯誤訊息，在遇到意外情況時給出。",
+  "501": "伺服器無法識別請求方法，或缺乏滿足請求的能力。",
+  "502": "伺服器充當閘道或代理，並從上游伺服器收到無效回應。",
+  "503": "伺服器目前不可用（因為過載或維護中）。",
+  "504": "伺服器充當閘道或代理，但未從上游伺服器收到及時回應。",
+  "505": "伺服器不支持請求中使用的 HTTP 協議版本。",
+  "506": "請求的透明內容協商導致循環引用。",
+  "507": "伺服器無法儲存完成請求所需的表示。",
+  "508": "伺服器在處理請求時檢測到無限循環。",
+  "510": "需要進一步擴展請求，伺服器才能完成它。",
+  "511": "用戶端需要身份驗證才能獲得網路訪問權限。"
+}
